### 2022-08-04 #4649 @editorialbot
[PRE REVIEW]: Vlasiator.jl: A Julia package for processing Vlasiator data
**Submitting author:** @henry2004y (<a href="http://orcid.org/0000-0003-4571-4501">Hongyang Zhou</a>)
**Repository:** https://github.com/henry2004y/Vlasiator.jl
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.9.25
