+++ src/vlsv/vlsvvariables.jl
@@ -83,7 +83,7 @@ const units_predefined = Dict(
 
 # Define derived parameters
 const variables_predefined = Dict(
-   :Bmag => function (meta, ids=UInt64[])
+   :Bmag => function (meta, ids=UInt[])
       rho_ = findfirst(endswith("rho"), meta.variable)
       ρ = readvariable(meta, meta.variable[rho_], ids)
       B = readvariable(meta, "vg_b_vol", ids)
@@ -92,7 +92,7 @@ const variables_predefined = Dict(
       _fillinnerBC!(Bmag, ρ)
       Bmag::Vector{Float32}
    end,
-   :Emag => function (meta, ids=UInt64[])
+   :Emag => function (meta, ids=UInt[])
       rho_ = findfirst(endswith("rho"), meta.variable)
       ρ = readvariable(meta, meta.variable[rho_], ids)
       E = readvariable(meta, "vg_e_vol", ids)
@@ -101,7 +101,7 @@ const variables_predefined = Dict(
       _fillinnerBC!(Emag, ρ)
       Emag::Vector{Float32}
    end,
-   :Vmag => function (meta, ids=UInt64[])
+   :Vmag => function (meta, ids=UInt[])
       rho_ = findfirst(endswith("rho"), meta.variable)
       ρ = readvariable(meta, meta.variable[rho_], ids)
       V = readvariable(meta, "proton/vg_v", ids)
@@ -110,7 +110,7 @@ const variables_predefined = Dict(
       _fillinnerBC!(Vmag, ρ)
       Vmag::Vector{Float32}
    end,
-   :Rhom => function (meta, ids=UInt64[])
+   :Rhom => function (meta, ids=UInt[])
       if hasvariable(meta, "vg_rhom")
          ρm = readvariable(meta, "vg_rhom", ids)
       elseif hasvariable(meta, "proton/vg_rho")
@@ -118,7 +118,7 @@ const variables_predefined = Dict(
       end
       ρm::Union{Vector{Float32}, Vector{Float64}}
    end,
-   :P => function (meta, ids=UInt64[]) # scalar pressure
+   :P => function (meta, ids=UInt[]) # scalar pressure
       if hasvariable(meta, "vg_pressure")
          P = readvariable(meta, "vg_pressure", ids)
       else
@@ -132,38 +132,38 @@ const variables_predefined = Dict(
       end
       P::Vector{Float32}
    end,
-   :VS => function (meta, ids=UInt64[]) # sound speed
+   :VS => function (meta, ids=UInt[]) # sound speed
       P = readvariable(meta, "P", ids)
       ρm = readvariable(meta, "Rhom", ids)
       _fillinnerBC!(ρm, ρm)
       VS = @. √( (P*5.0f0/3.0f0) / ρm )
       VS::Union{Vector{Float32}, Vector{Float64}}
    end,
-   :VA => function (meta, ids=UInt64[]) # Alfvén speed
+   :VA => function (meta, ids=UInt[]) # Alfvén speed
       ρm = readvariable(meta, "Rhom", ids)
       _fillinnerBC!(ρm, ρm)
       Bmag = readvariable(meta, "Bmag", ids)
       VA = @. Bmag / √(ρm*μ₀)
       VA::Union{Vector{Float32}, Vector{Float64}}
    end,
-   :MA => function (meta, ids=UInt64[]) # Alfvén Mach number
+   :MA => function (meta, ids=UInt[]) # Alfvén Mach number
       V = readvariable(meta, "Vmag", ids)
       VA = readvariable(meta, "VA", ids)
       MA = V ./ VA
[…]
