### 2022-09-03 a2d0e37
First attempt for regex parsing
