+++ src/vlsv/vlsvreader.jl
@@ -29,7 +29,7 @@ struct MetaVLSV
    dir::String
    name::String
    fid::IOStream
-   footer::EzXML.Node
+   footer::String
    variable::Vector{String}
    "mapping of unsorted cell ID to ordering"
    celldict::Dict{UInt, Int}
@@ -79,61 +79,9 @@ end
    # Obtain the offset of the XML footer
    offset = read(fid, UInt)
    seek(fid, offset)
-   footer = read(fid, String) |> parsexml |> root
+   footer = read(fid, String)
 end
 
-
-"Return size and type information for the inquired parameter/data with `name`."
-function getObjInfo(footer::EzXML.Node, name::String, tag::String, attr::String)
-   local arraysize, datasize, datatype, vectorsize, variable_offset
-   isFound = false
-
-   for var in findall("//$tag", footer)
-      if var[attr] == String(name)
-         arraysize = parse(Int, var["arraysize"])
-         datasize = parse(Int, var["datasize"])
-         datatype = var["datatype"]
-         vectorsize = parse(Int, var["vectorsize"])
-         variable_offset = parse(Int, nodecontent(var))
-         isFound = true
-         break
-      end
-   end
-
-   isFound || throw(ArgumentError("unknown variable $name"))
-
-   T::Type =
-      if datatype == "float"
-         datasize == 4 ? Float32 : Float64
-      elseif datatype == "int"
-         datasize == 4 ? Int32 : Int
-      elseif datatype == "uint"
-         datasize == 4 ? UInt32 : UInt
-      end
-
-   T, variable_offset, arraysize, datasize, vectorsize
-end
-
-"Return vector of `name` from the VLSV file with `footer` associated with stream `fid`."
-function readvector(fid::IOStream, footer::EzXML.Node, name::String, tag::String)
-   T, offset, asize, dsize, vsize = getObjInfo(footer, name, tag, "name")
-
-   if Sys.total_memory() > 8*asize*vsize*dsize
-      w = vsize == 1 ?
-         Vector{T}(undef, asize) :
-         Array{T,2}(undef, vsize, asize)
-      seek(fid, offset)
-      read!(fid, w)
-   else
-      @warn "Large array detected. Using memory-mapped I/O!" maxlog=1
-      a = mmap(fid, Vector{UInt8}, dsize*vsize*asize, offset)
-      w = vsize == 1 ?
-         reinterpret(T, a) :
-         reshape(reinterpret(T, a), vsize, asize)
-   end
-
-   w
-end
 
 """
     load(file::AbstractString)) -> MetaVLSV
@@ -172,38 +120,23 @@ function load(file::AbstractString)
    vmin = [0.0, 0.0, 0.0]
    vmax = [1.0, 1.0, 1.0]
 
-   for varinfo in findall("//BLOCKIDS", footer)
+   # Only support VLSV 5.0+ file
+   pattern = r"<BLOCKIDS.*name=\"(?<name>.*?)\".*</BLOCKIDS>"
+   ms = eachmatch(pattern, footer)
 
-      if haskey(varinfo, "name")
-         # VLSV 5.0 file with bounding box
-         popname = varinfo["name"]
+   for m in ms
+      popname = m[:name]
 
-         vbox = readmesh(fid, footer, popname, "MESH_BBOX")::Vector{UInt}
[…]
