### 2022-09-04 bd354e8
Collect nodes into struct
