+++ src/vlsv/vlsvreader.jl
@@ -24,17 +24,21 @@ struct VarInfo
    unitConversion::String
 end
 
+struct NodeVLSV
+   var::Vector{EzXML.Node}
+   param::Vector{EzXML.Node}
+   cellwithVDF::Vector{EzXML.Node}
+   cellblocks::Vector{EzXML.Node}
+   blockvar::Vector{EzXML.Node}
+   blockid::Vector{EzXML.Node}
+end
+
 "VLSV meta data."
 struct MetaVLSV
    dir::String
    name::String
    fid::IOStream
-   nodevar::Vector{EzXML.Node}
-   nodeparam::Vector{EzXML.Node}
-   nodecellwithVDF::Vector{EzXML.Node}
-   nodecellblocks::Vector{EzXML.Node}
-   nodeblockvar::Vector{EzXML.Node}
-   nodeblockid::Vector{EzXML.Node}
+   nodeVLSV::NodeVLSV
    variable::Vector{String}
    "mapping of unsorted cell ID to ordering"
    celldict::Dict{UInt, Int}
@@ -207,7 +211,10 @@ function load(file::AbstractString)
    nodeblockvar = findall("//BLOCKVARIABLE", footer)
    nodeblockid = findall("//BLOCKIDS", footer)
 
-   cellid = getcellid(fid, nodevar)
+   n = NodeVLSV(nodevar, nodeparam, nodecellwithVDF, nodecellblocks, nodeblockvar,
+      nodeblockid)
+
+   cellid = getcellid(fid, n.var)
 
    cellindex = sortperm(cellid)
 
@@ -225,7 +232,7 @@ function load(file::AbstractString)
    vmin = [0.0, 0.0, 0.0]
    vmax = [1.0, 1.0, 1.0]
 
-   for node in nodeblockid
+   for node in n.blockid
       if haskey(node, "name")
          # VLSV 5.0 file with bounding box
          popname = node["name"]
@@ -240,13 +247,13 @@ function load(file::AbstractString)
       else
          popname = "avgs"
          # In VLSV before 5.0 the mesh is defined with parameters.
-         if "vxblocks_ini" in getindex.(nodeparam, "name")
+         if "vxblocks_ini" in getindex.(n.param, "name")
             vblocks_str = ("vxblocks_ini", "vyblocks_ini", "vzblocks_ini")
             vmin_str = ("vxmin", "vymin", "vzmin")
             vmax_str = ("vxmax", "vymax", "vzmax")
-            vblocks .= [readparameter(fid, nodeparam, vblocks_str[i]) for i in 1:3]
-            vmin .= [readparameter(fid, nodeparam, vmin_str[i]) for i in 1:3]
-            vmax .= [readparameter(fid, nodeparam, vmax_str[i]) for i in 1:3]
+            vblocks .= [readparameter(fid, n.param, vblocks_str[i]) for i in 1:3]
+            vmin .= [readparameter(fid, n.param, vmin_str[i]) for i in 1:3]
+            vmax .= [readparameter(fid, n.param, vmax_str[i]) for i in 1:3]
             dv = ntuple(i -> (vmax[i] - vmin[i]) / vblocks[i] / vblock_size[i], Val(3))
          else
             error("File not written by Vlasiator!")
@@ -269,10 +276,10 @@ function load(file::AbstractString)
       meshes[popname] = popVMesh
    end
 
-   if hasname(nodeparam, "time") # Vlasiator 5.0+
-      timesim = readparameter(fid, nodeparam, "time")::Float64
-   elseif hasname(nodeparam, "t")
-      timesim = readparameter(fid, nodeparam, "t")::Float64
+   if hasname(n.param, "time") # Vlasiator 5.0+
[…]
