### 2022-09-04 a82d25c
Improved node splitting
