### 2022-09-04 d96e0d9
Refactor reading (#98) [skip ci]
* Replace a single footer with fine-grained nodes
* Remove unnecessary precompile
* Fix potential bug in multipop vcell reading
* Collect nodes into struct
* Two sweep for node splitting
* Rename function
* Bump patch
