+++ src/vlsv/vlsvreader.jl
@@ -2,6 +2,9 @@
 
 include("vlsvvariables.jl")
 
+const NodeVector = SubArray{EzXML.Node, 1, Vector{EzXML.Node}, Tuple{UnitRange{Int64}},
+   true}
+
 "Velocity mesh information."
 struct VMeshInfo
    "number of velocity blocks"
@@ -24,12 +27,21 @@ struct VarInfo
    unitConversion::String
 end
 
+struct NodeVLSV
+   var::NodeVector
+   param::NodeVector
+   cellwithVDF::NodeVector
+   cellblocks::NodeVector
+   blockvar::NodeVector
+   blockid::NodeVector
+end
+
 "VLSV meta data."
 struct MetaVLSV
    dir::String
    name::String
    fid::IOStream
-   footer::EzXML.Node
+   nodeVLSV::NodeVLSV
    variable::Vector{String}
    "mapping of unsorted cell ID to ordering"
    celldict::Dict{UInt, Int}
@@ -76,32 +88,14 @@ end
    # First 8 bytes indicate big-endian or else
    endian_offset = 8
    seek(fid, endian_offset)
-   # Obtain the offset of the XML footer
+   # Obtain the offset of the XML
    offset = read(fid, UInt)
    seek(fid, offset)
    footer = read(fid, String) |> parsexml |> root
 end
 
 
-"Return size and type information for the inquired parameter/data with `name`."
-function getObjInfo(footer::EzXML.Node, name::String, tag::String, attr::String)
-   local arraysize, datasize, datatype, vectorsize, variable_offset
-   isFound = false
-
-   for var in findall("//$tag", footer)
-      if var[attr] == String(name)
-         arraysize = parse(Int, var["arraysize"])
-         datasize = parse(Int, var["datasize"])
-         datatype = var["datatype"]
-         vectorsize = parse(Int, var["vectorsize"])
-         variable_offset = parse(Int, nodecontent(var))
-         isFound = true
-         break
-      end
-   end
-
-   isFound || throw(ArgumentError("unknown variable $name"))
-
+@inline function getdatatype(datatype::String, datasize::Int)
    T::Type =
       if datatype == "float"
          datasize == 4 ? Float32 : Float64
@@ -109,14 +103,82 @@ function getObjInfo(footer::EzXML.Node, name::String, tag::String, attr::String)
          datasize == 4 ? Int32 : Int
       elseif datatype == "uint"
          datasize == 4 ? UInt32 : UInt
+      else
+         throw(ArgumentError("unknown type $datatype"))
       end
-
-   T, variable_offset, arraysize, datasize, vectorsize
 end
 
-"Return vector of `name` from the VLSV file with `footer` associated with stream `fid`."
-function readvector(fid::IOStream, footer::EzXML.Node, name::String, tag::String)
-   T, offset, asize, dsize, vsize = getObjInfo(footer, name, tag, "name")
+function getvarinfo(nodevar::AbstractVector{EzXML.Node}, name::String)
+   local arraysize, datasize, datatype, vectorsize, offset
+   isFound = false
+
+   for var in nodevar
+      if var["name"] == name
+         arraysize = parse(Int, var["arraysize"])
+         datasize = parse(Int, var["datasize"])
+         datatype = var["datatype"]
+         vectorsize = parse(Int, var["vectorsize"])
+         offset = parse(Int, nodecontent(var))
+         isFound = true
[…]
