+++ src/Vlasiator.jl
@@ -40,31 +40,13 @@ export
    # fluxfunction
    compute_flux_function, find_reconnection_points
 
-precompile(load, (String,))
 precompile(readvariable, (MetaVLSV, String, Vector{UInt}))
 precompile(readvariable, (MetaVLSV, String, Int))
-precompile(readvariable, (MetaVLSV, String))
 precompile(readvariable, (MetaVLSV, String, UnitRange{Int}))
-precompile(readvariable, (MetaVLSV, String, Bool))
 precompile(readvcells, (MetaVLSV, Int))
 precompile(readvcells, (MetaVLSV, UInt))
-precompile(readlog, (String,))
-precompile(readmesh, (IOStream, EzXML.Node, String, String))
-precompile(readparameter, (MetaVLSV, String))
-precompile(readparameter, (IOStream, EzXML.Node, String))
-precompile(readvariablemeta, (MetaVLSV, String))
-precompile(readvector, (IOStream, EzXML.Node, String, String))
 precompile(getcell, (MetaVLSV, Vector{Float64}))
-precompile(getObjInfo, (EzXML.Node, String, String, String))
-precompile(get_axis, (Vlasiator.PlotArgs,))
-precompile(getcellwithvdf, (MetaVLSV,))
-precompile(getdata2d, (MetaVLSV, String))
-precompile(getfooter, (IOStream,))
-precompile(hasname, (EzXML.Node, String, String))
-precompile(hasparameter, (MetaVLSV, String))
-precompile(hasvariable, (MetaVLSV, String))
-precompile(prep2d, (MetaVLSV, String))
-precompile(set_args, (MetaVLSV, String, Vlasiator.AxisUnit))
+precompile(prep2d, (MetaVLSV, String, Int))
 precompile(searchsorted, (Vector{UInt}, Int))
 
 function __init__()
