### 2022-09-05 64a6ff3
One sweep for node splitting
