### 2022-09-21 b764c9c
Add mmap test
