### 2022-09-24 03ac4d5
More strict type
