### 2022-09-24 938d632
Refactor readvector; More strict types
