### 2022-09-25 229f330
Add compact for SnoopPrecompile
