+++ Project.toml
@@ -27,6 +27,7 @@ LazyGrids = "0.4"
 PyPlot = "2.9"
 RecipesBase = "1.1"
 Requires = "1.1"
+SnoopPrecompile = "1.0"
 StaticArrays = "1.2"
 WriteVTK = "1.9"
 julia = "1.7"
