### 2022-09-26 cb66423
Revert "Refactor mmap" and fix return types
This reverts commit 8345d1115b9845dabd283eb0b315d7b063a75324.
