### 2022-09-26 b2cca96
Use mmap for CellID seems fine
