### 2022-09-30 #4649 @pdebuyl
@StanczakDominik @akaptano would you review the article "Vlasiator.jl: A Julia package for processing Vlasiator data" for the Journal of Open Source Software?
