### 2022-10-02 29e5d5d
Do not store initfile for precompilation
