### 2022-10-11 ff23920
Working native XML reader
