+++ Project.toml
@@ -1,7 +1,7 @@
 name = "Vlasiator"
 uuid = "7d2ba682-ad6e-4e20-80d9-3f2d4a610bb4"
 authors = ["Hongyang Zhou <hyzhou@umich.edu>"]
-version = "0.9.29"
+version = "0.9.30"
 
 [deps]
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
@@ -10,6 +10,7 @@ LaTeXStrings = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f"
 LazyGrids = "7031d0ef-c40d-4431-b2f8-61a8d2f650db"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Mmap = "a63ad114-7e13-5084-954f-fe012c677804"
+Parsers = "69de0a69-1ddd-5017-9359-2bf0b02dc9f0"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 PyPlot = "d330b81b-6aea-500a-939a-2ce795aea3ee"
 REPL = "3fa0cd96-eef1-5676-8a61-b3b8758bbffb"
@@ -24,6 +25,7 @@ WriteVTK = "64499a7a-5c06-52f2-abe2-ccb03c286192"
 EzXML = "1.1"
 LaTeXStrings = "1.2"
 LazyGrids = "0.4"
+Parsers = "2.4"
 PyPlot = "2.9"
 RecipesBase = "1.1"
 Requires = "1.1"
