### 2022-10-18 7e965a5
Reduce duplicate memory access for derived vars; Add type annotations
