+++ src/vlsv/vlsvvariables.jl
@@ -83,34 +83,40 @@ const units_predefined = Dict(
 
 # Define derived parameters
 const variables_predefined = Dict(
-   :Bmag => function (meta, ids=UInt[])
-      rho_ = findfirst(endswith("rho"), meta.variable)
-      ρ = readvariable(meta, meta.variable[rho_], ids)
+   :Bmag => function (meta::MetaVLSV, ids::Vector{<:Integer}=UInt[])
+      ρ = _readrho(meta, ids)
       B = readvariable(meta, "vg_b_vol", ids)
       Bmag = sum(x -> x*x, B, dims=1) |> vec
       @. Bmag = √(Bmag)
       _fillinnerBC!(Bmag, ρ)
       Bmag::Vector{Float32}
    end,
-   :Emag => function (meta, ids=UInt[])
-      rho_ = findfirst(endswith("rho"), meta.variable)
-      ρ = readvariable(meta, meta.variable[rho_], ids)
+   :Emag => function (meta::MetaVLSV, ids::Vector{<:Integer}=UInt[])
+      ρ = _readrho(meta, ids)
       E = readvariable(meta, "vg_e_vol", ids)
       Emag = sum(x -> x*x, E, dims=1) |> vec
       @. Emag = √(Emag)
       _fillinnerBC!(Emag, ρ)
       Emag::Vector{Float32}
    end,
-   :Vmag => function (meta, ids=UInt[])
-      rho_ = findfirst(endswith("rho"), meta.variable)
-      ρ = readvariable(meta, meta.variable[rho_], ids)
+   :Vmag => function (meta::MetaVLSV, ids::Vector{<:Integer}=UInt[])
+      ρ = _readrho(meta, ids)
       V = readvariable(meta, "proton/vg_v", ids)
       Vmag = sum(x -> x*x, V, dims=1) |> vec
       @. Vmag = √(Vmag)
       _fillinnerBC!(Vmag, ρ)
       Vmag::Vector{Float32}
    end,
-   :Rhom => function (meta, ids=UInt[])
+   :Bhat => function (meta::MetaVLSV, ids::Vector{<:Integer}=UInt[])
+      ρ = _readrho(meta, ids)
+      Bhat = readvariable(meta, "vg_b_vol", ids)
+      Bmag = sum(x -> x*x, Bhat, dims=1) |> vec
+      @. Bmag = √(Bmag)
+      Bhat ./= Bmag'
+      _fillinnerBC!(Bhat, ρ)
+      Bhat::Matrix{Float32}
+   end,
+   :Rhom => function (meta::MetaVLSV, ids::Vector{<:Integer}=UInt[])
       if hasvariable(meta, "vg_rhom")
          ρm = readvariable(meta, "vg_rhom", ids)
       elseif hasvariable(meta, "proton/vg_rho")
@@ -118,7 +124,7 @@ const variables_predefined = Dict(
       end
       ρm::Union{Vector{Float32}, Vector{Float64}}
    end,
-   :P => function (meta, ids=UInt[]) # scalar pressure
+   :P => function (meta::MetaVLSV, ids::Vector{<:Integer}=UInt[]) # scalar pressure
       if hasvariable(meta, "vg_pressure")
          P = readvariable(meta, "vg_pressure", ids)
       else
@@ -132,40 +138,40 @@ const variables_predefined = Dict(
       end
       P::Vector{Float32}
    end,
-   :VS => function (meta, ids=UInt[]) # sound speed
+   :VS => function (meta::MetaVLSV, ids::Vector{<:Integer}=UInt[]) # sound speed
       P = readvariable(meta, "P", ids)
       ρm = readvariable(meta, "Rhom", ids)
[…]
