### 2022-10-20 c155df9
Add colors to repl show
