### 2022-10-21 6c9e8f5
Annotate arg types [skip ci]
