+++ src/vlsv/vlsvreader.jl
@@ -6,8 +6,8 @@ const NodeVector = SubArray{EzXML.Node, 1, Vector{EzXML.Node}, Tuple{UnitRange{I
 "Velocity mesh information."
 struct VMeshInfo
    "number of velocity blocks"
-   vblocks::NTuple{3, UInt}
-   vblock_size::NTuple{3, UInt}
+   vblocks::NTuple{3, Int}
+   vblock_size::NTuple{3, Int}
    vmin::NTuple{3, Float64}
    vmax::NTuple{3, Float64}
    dv::NTuple{3, Float64}
@@ -42,7 +42,7 @@ struct MetaVLSV
    nodeVLSV::NodeVLSV
    variable::Vector{String}
    "mapping of unsorted cell ID to ordering"
-   celldict::Dict{UInt, Int}
+   celldict::Dict{Int, Int}
    "ordered sequence indexes of raw cell IDs"
    cellindex::Vector{Int}
    time::Float64
@@ -93,7 +93,7 @@ end
    endian_offset = 8
    seek(fid, endian_offset)
    # Obtain the offset of the XML
-   offset = read(fid, UInt)
+   offset = read(fid, Int)
    seek(fid, offset)
    footer = read(fid, String) |> parsexml |> root
 end
@@ -269,8 +269,8 @@ function load(file::AbstractString)
    # Find all species by the BLOCKIDS tag
    species = String[]
 
-   vblocks = UInt[0, 0, 0]
-   vblock_size = UInt[4, 4, 4]
+   vblocks = [0, 0, 0]
+   vblock_size = [4, 4, 4]
    vmin = [0.0, 0.0, 0.0]
    vmax = [1.0, 1.0, 1.0]
 
@@ -350,7 +350,7 @@ function load(f::Function, file::AbstractString)
    end
 end
 
-function getmaxrefinement(cellid::AbstractArray{UInt, 1}, ncells::NTuple{3, UInt})
+function getmaxrefinement(cellid::Vector{Int}, ncells::NTuple{3, Int})
    ncell = prod(ncells)
    maxamr, cid = 0, ncell
    while @inbounds cid < maximum(cellid)
@@ -427,7 +427,7 @@ function readmesh(fid::IOStream, footer::EzXML.Node)
    node = findfirst("//MESH_BBOX", footer)
    offset = Parsers.parse(Int, nodecontent(node))
 
-   bbox = Vector{UInt}(undef, 6)
+   bbox = Vector{Int}(undef, 6)
    seek(fid, offset)
    read!(fid, bbox)
 
@@ -447,7 +447,7 @@ end
 function readvmesh(fid::IOStream, footer::EzXML.Node, species::String)
    ns = findall("//MESH_BBOX", footer)
 
-   bbox = Vector{UInt}(undef, 6)
+   bbox = Vector{Int}(undef, 6)
 
    for i in eachindex(ns)[3:end]
       if ns[i]["mesh"] == species
@@ -513,12 +513,12 @@ function readvariable(meta::MetaVLSV, var::String, sorted::Bool=true, usemmap::B
 end
 
 """
-    readvariable(meta::MetaVLSV, var::String, ids::Vector{<:Integer}) -> Array
+    readvariable(meta::MetaVLSV, var::String, ids::Vector{Int}) -> Array
 
 Read variable `var` in a collection of cells `ids` associated with `meta`. if `ids` is
 empty, return the whole sorted array of `var`.
 """
-function readvariable(meta::MetaVLSV, var::String, ids::Vector{<:Integer})::Array
+function readvariable(meta::MetaVLSV, var::String, ids::Vector{Int})::Array
    startswith(var, "fg_") && error("Currently does not support reading fsgrid!")
 
    if (local symvar = Symbol(var)) in keys(variables_predefined)
[…]
