+++ src/vlsv/vlsvvariables.jl
@@ -83,7 +83,7 @@ const units_predefined = Dict(
 
 # Define derived parameters
 const variables_predefined = Dict(
-   :Bmag => function (meta::MetaVLSV, ids::Vector{<:Integer}=UInt[])
+   :Bmag => function (meta::MetaVLSV, ids::Vector{Int}=Int[])
       ρ = _readrho(meta, ids)
       B = readvariable(meta, "vg_b_vol", ids)
       Bmag = sum(x -> x*x, B, dims=1) |> vec
@@ -91,7 +91,7 @@ const variables_predefined = Dict(
       _fillinnerBC!(Bmag, ρ)
       Bmag::Vector{Float32}
    end,
-   :Emag => function (meta::MetaVLSV, ids::Vector{<:Integer}=UInt[])
+   :Emag => function (meta::MetaVLSV, ids::Vector{Int}=Int[])
       ρ = _readrho(meta, ids)
       E = readvariable(meta, "vg_e_vol", ids)
       Emag = sum(x -> x*x, E, dims=1) |> vec
@@ -99,7 +99,7 @@ const variables_predefined = Dict(
       _fillinnerBC!(Emag, ρ)
       Emag::Vector{Float32}
    end,
-   :Vmag => function (meta::MetaVLSV, ids::Vector{<:Integer}=UInt[])
+   :Vmag => function (meta::MetaVLSV, ids::Vector{Int}=Int[])
       ρ = _readrho(meta, ids)
       V = readvariable(meta, "proton/vg_v", ids)
       Vmag = sum(x -> x*x, V, dims=1) |> vec
@@ -107,7 +107,7 @@ const variables_predefined = Dict(
       _fillinnerBC!(Vmag, ρ)
       Vmag::Vector{Float32}
    end,
-   :Bhat => function (meta::MetaVLSV, ids::Vector{<:Integer}=UInt[])
+   :Bhat => function (meta::MetaVLSV, ids::Vector{Int}=Int[])
       ρ = _readrho(meta, ids)
       Bhat = readvariable(meta, "vg_b_vol", ids)
       Bmag = sum(x -> x*x, Bhat, dims=1) |> vec
@@ -116,7 +116,7 @@ const variables_predefined = Dict(
       _fillinnerBC!(Bhat, ρ)
       Bhat::Matrix{Float32}
    end,
-   :Rhom => function (meta::MetaVLSV, ids::Vector{<:Integer}=UInt[])
+   :Rhom => function (meta::MetaVLSV, ids::Vector{Int}=Int[])
       if hasvariable(meta, "vg_rhom")
          ρm = readvariable(meta, "vg_rhom", ids)
       elseif hasvariable(meta, "proton/vg_rho")
@@ -124,7 +124,7 @@ const variables_predefined = Dict(
       end
       ρm::Union{Vector{Float32}, Vector{Float64}}
    end,
-   :P => function (meta::MetaVLSV, ids::Vector{<:Integer}=UInt[]) # scalar pressure
+   :P => function (meta::MetaVLSV, ids::Vector{Int}=Int[]) # scalar pressure
       if hasvariable(meta, "vg_pressure")
          P = readvariable(meta, "vg_pressure", ids)
       else
@@ -138,38 +138,38 @@ const variables_predefined = Dict(
       end
       P::Vector{Float32}
    end,
-   :VS => function (meta::MetaVLSV, ids::Vector{<:Integer}=UInt[]) # sound speed
+   :VS => function (meta::MetaVLSV, ids::Vector{Int}=Int[]) # sound speed
       P = readvariable(meta, "P", ids)
       ρm = readvariable(meta, "Rhom", ids)
[…]
