### 2022-10-27 8d931b3
Optimize mem for maxamronly
