+++ test/runtests.jl
@@ -314,9 +314,8 @@ end
          @test sha_str == "8a2bb0a15c5dcc329f88821036df840a86eef9d5"
 
          # Selected region
-         write_vtk(meta, vars=["proton/vg_rho"], box=
-            [meta.coordmin[1], meta.coordmax[1], 0, meta.coordmax[2], 0, meta.coordmax[3]],
-            maxamronly=true)
+         write_vtk(meta, vars=["proton/vg_rho"], maxamronly=true, box=
+            [meta.coordmin[1], meta.coordmax[1], 0, meta.coordmax[2], 0, meta.coordmax[3]])
          sha_str = bytes2hex(open(sha1, "bulk.amr.vti"))
          @test sha_str == "50e01f51ec7e16a1a57e794eab8545eeeda4e2b6"
 
