+++ src/vlsv/vlsvutility.jl
@@ -23,11 +23,8 @@ function getcell(meta::MetaVLSV, loc::AbstractVector{<:AbstractFloat})
       haskey(celldict, cid) && break
 
       ncells_lowerlevel += (8^ilevel)*ncell
-
       ratio = 2^(ilevel+1)
-
       indices = ntuple(i -> floor(Int, (loc[i] - coordmin[i]) / dcoord[i] * ratio), Val(3))
-
       cid = ncells_lowerlevel + indices[1] +
          ratio*ncells[1]*indices[2] + ratio^2*ncells[1]*ncells[2]*indices[3] + 1
    end
@@ -955,7 +952,7 @@ fillmesh(meta::MetaVLSV, vars::String) = fillmesh(meta, [vars])
 
 """
     fillmesh(meta::MetaVLSV, vars::Vector{String};
-       skipghosttype=true, verbose=false) -> celldata, vtkGhostType
+       skipghosttype=true, maxamronly=false, verbose=false) -> celldata, vtkGhostType
 
 Fill the DCCRG mesh with quantity of `vars` on all refinement levels.
 # Return arguments
@@ -963,7 +960,7 @@ Fill the DCCRG mesh with quantity of `vars` on all refinement levels.
 - `vtkGhostType::Array{UInt8}`: cell status (to be completed!).
 """
 function fillmesh(meta::MetaVLSV, vars::Vector{String};
-   skipghosttype::Bool=true, verbose::Bool=false)
+   skipghosttype::Bool=true, maxamronly::Bool=false, verbose::Bool=false)
 
    (;maxamr, fid, nodeVLSV, ncells, celldict) = meta
 
@@ -982,12 +979,19 @@ function fillmesh(meta::MetaVLSV, vars::Vector{String};
       if T[i] == Float64 Tout[i] = Float32 end
    end
 
-   @inbounds celldata =
+   @inbounds celldata = if !maxamronly
       [[zeros(Tout[iv], vsize[iv], ncells[1] << i, ncells[2] << i, ncells[3] << i)
       for i = 0:maxamr] for iv in 1:nv]
+   else
+      [[zeros(Tout[iv], vsize[iv],
+      ncells[1] << maxamr, ncells[2] << maxamr, ncells[3] << maxamr)] for iv in 1:nv]
+   end
 
-   @inbounds vtkGhostType =
+   @inbounds vtkGhostType = if !skipghosttype
       [zeros(UInt8, ncells[1] << i, ncells[2] << i, ncells[3] << i) for i = 0:maxamr]
+   else
+      [UInt8[]]
+   end
 
    if maxamr == 0
       @inbounds for iv = 1:nv
@@ -1007,7 +1011,7 @@ function fillmesh(meta::MetaVLSV, vars::Vector{String};
       idfirst_ = searchsortedfirst(cellidsorted, nLow+1)
       idlast_  = searchsortedlast(cellidsorted, nHigh)
 
-      ids = cellidsorted[idfirst_:idlast_]
+      ids = @view cellidsorted[idfirst_:idlast_]
 
       if !skipghosttype
          # Mark non-existing cells due to refinement
@@ -1029,8 +1033,11 @@ function fillmesh(meta::MetaVLSV, vars::Vector{String};
             a = mmap(fid, Vector{UInt8}, sizeof(T[iv])*vsize[iv]*arraysize[iv], offset[iv])
             dataRaw = reshape(reinterpret(T[iv], a), vsize[iv], arraysize[iv])
             data = @view dataRaw[:,rOffsetsRaw]
-
-            _fillcell!(ilvl, maxamr, ids, ix, iy, iz, celldata[iv], data)
+            if !maxamronly
[…]
