### 2022-10-27 4c4536d
Successor of #82; handle #101
