+++ src/vlsv/vlsvvariables.jl
@@ -77,8 +77,8 @@ const units_predefined = Dict(
    :BetaStar => ("",L"$\beta^\ast$", ""),
    :Gyroperiod => ("s", L"$T_{gyro}$", "s"),
    :PlasmaPeriod => ("s", L"$T_{plasma}$", "s"),
-   :Gyrofrequency => ("/s", L"$\omega_{g}$", L"s^{-1}"),
-   :Omegap => ("/s", L"$\omega_{p}$", L"s^{-1}"),
+   :Gyrofrequency => ("rad/s", L"$\omega_{g}$", L"\mathrm{rad}/\mathrm{s}"),
+   :Omegap => ("rad/s", L"$\omega_{p}$", L"\mathrm{rad}/\mathrm{s}"),
 )
 
 # Define derived parameters
@@ -174,16 +174,12 @@ const variables_predefined = Dict(
       b = readvariable(meta, "Bhat", ids)
       Vperp = Vector{Float32}(undef, size(V, 2))
 
-      function _computeVperp!()
-         # Avoid sqrt of negative values, but does not guarantee orthogonality.
-         @inbounds @views for i in eachindex(Vperp)
-            vb = V[:,i] ⋅ b[:,i]
-            Vpar = SVector{3,Float32}(vb*b[1,i], vb*b[2,i], vb*b[3,i])
-            Vperp[i] = norm(V[:,i] - Vpar)
-         end
+      # Avoid sqrt of negative values, but does not guarantee orthogonality.
+      @inbounds @views for i in eachindex(Vperp)
+         vb = V[:,i] ⋅ b[:,i]
+         Vpar = SVector{3,Float32}(vb*b[1,i], vb*b[2,i], vb*b[3,i])
+         Vperp[i] = norm(V[:,i] - Vpar)
       end
-
-      _computeVperp!()
 
       Vperp
    end,
@@ -191,13 +187,13 @@ const variables_predefined = Dict(
       P = readvariable(meta, "P", ids)
       n = readvariable(meta, "n", ids)
       _fillinnerBC!(n, n)
-      T = @. P / (n*kB)
-      T::Vector{Float64}
+      T = @. P / (n*Float32(kB))
+      T::Vector{Float32}
    end,
    :Vth => function (meta::MetaVLSV, ids::Vector{Int}=Int[]) # thermal velocity
       T = readvariable(meta, "T", ids)
-      Vth = @. √(3 * kB * T / mᵢ) # assume proton
-      Vth::Vector{Float64}
+      Vth = @. √(3 * Float32(kB) * T / Float32(mᵢ)) # assume proton
+      Vth::Vector{Float32}
    end,
    :Ppar => function (meta::MetaVLSV, ids::Vector{Int}=Int[]) # P component ∥ B
       P = readvariable(meta, "Protated", ids)
@@ -211,29 +207,51 @@ const variables_predefined = Dict(
       P = readvariable(meta, "Protated", ids)
       n = readvariable(meta, "n", ids)
       _fillinnerBC!(n, n)
-      Tpar = @. P[3,3,:] / (n*kB)
-      Tpar::Vector{Float64}
+      Tpar = @. P[3,3,:] / (n*Float32(kB))
+      Tpar::Vector{Float32}
    end,
    :Tperp => function (meta::MetaVLSV, ids::Vector{Int}=Int[]) # scalar T component ⟂ B
       P = readvariable(meta, "Protated", ids)
       n = readvariable(meta, "n", ids)
       _fillinnerBC!(n, n)
[…]
