### 2022-11-02 bb92809
Add Jpar, Jperp to table
