### 2022-11-03 822bbec
Improve set_colorbar
