### 2022-11-08 423b714
Add refineslice for vector vars
