### 2022-11-09 89cdd1c
Support streamplot and quiver for amr
