+++ versions.js
@@ -11,5 +11,5 @@ var DOC_VERSIONS = [
   "v0.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.9.30";
+var DOCUMENTER_NEWEST = "v0.9.32";
 var DOCUMENTER_STABLE = "stable";
