### 2022-11-17 1cb8ffe
Reduce 2d plotting duplication
