### 2022-11-23 henry2004y/Vlasiator.jl#114 @ranocha
JOSS review: test failure on Julia 1.8
On Julia v1.8.2, I get

```julia
(@v1.8) pkg> activate .

(Vlasiator_jl) pkg> add Vlasiator
...

(Vlasiator_jl) pkg> test Vlasiator
...
  Downloaded artifact: testdata
┌ Warning: Error requiring `PyPlot` from `Vlasiator`
│   exception =
│    LoadError: PyError ($(Expr(:escape, :(ccall(#= ~/.julia/packages/PyCall/ygXW2/src/pyfncall.jl:43 =# @pysym(:PyObject_Call), PyPtr, (PyPtr, PyPtr, PyPtr), o, pyargsptr, kw))))) <class 'KeyError'>
│    KeyError('Unrecognized key "pcolor.shading" for group "pcolor" and name "shading"')
│      File "/usr/lib/python3/dist-packages/matplotlib/__init__.py", line 1115, in rc
│        raise KeyError(('Unrecognized key "%s" for group "%s" and '
│
│    Stacktrace:
│      [1] pyerr_check
│        @ ~/.julia/packages/PyCall/ygXW2/src/exception.jl:62 [inlined]
│      [2] pyerr_check
│        @ ~/.julia/packages/PyCall/ygXW2/src/exception.jl:66 [inlined]
│      [3] _handle_error(msg::String)
│        @ PyCall ~/.julia/packages/PyCall/ygXW2/src/exception.jl:83
│      [4] macro expansion
│        @ ~/.julia/packages/PyCall/ygXW2/src/exception.jl:97 [inlined]
│      [5] #107
│        @ ~/.julia/packages/PyCall/ygXW2/src/pyfncall.jl:43 [inlined]
│      [6] disable_sigint
│        @ ./c.jl:473 [inlined]
│      [7] __pycall!
│        @ ~/.julia/packages/PyCall/ygXW2/src/pyfncall.jl:42 [inlined]
│      [8] _pycall!(ret::PyCall.PyObject, o::PyCall.PyObject, args::Tuple{String}, nargs::Int64, kw::PyCall.PyObject)
│        @ PyCall ~/.julia/packages/PyCall/ygXW2/src/pyfncall.jl:29
│      [9] _pycall!(ret::PyCall.PyObject, o::PyCall.PyObject, args::Tuple{String}, kwargs::Base.Pairs{Symbol, String, Tuple{Symbol}, NamedTuple{(:shading,), Tuple{String}}})
│        @ PyCall ~/.julia/packages/PyCall/ygXW2/src/pyfncall.jl:11
│     [10] #_#114
│        @ ~/.julia/packages/PyCall/ygXW2/src/pyfncall.jl:86 [inlined]
│     [11] top-level scope
│        @ ~/.julia/packages/Vlasiator/oOdeL/src/plot/pyplot.jl:15
│     [12] include(mod::Module, _path::String)
│        @ Base ./Base.jl:419
│     [13] include(x::String)
│        @ Vlasiator ~/.julia/packages/Vlasiator/oOdeL/src/Vlasiator.jl:1
│     [14] macro expansion
│        @ ~/.julia/packages/Requires/Z8rfN/src/Requires.jl:40 [inlined]
│     [15] top-level scope
│        @ ~/.julia/packages/Vlasiator/oOdeL/src/Vlasiator.jl:54
│     [16] eval
│        @ ./boot.jl:368 [inlined]
│     [17] eval
│        @ ~/.julia/packages/Vlasiator/oOdeL/src/Vlasiator.jl:1 [inlined]
│     [18] (::Vlasiator.var"#231#237")()
│        @ Vlasiator ~/.julia/packages/Requires/Z8rfN/src/require.jl:101
│     [19] macro expansion
│        @ ./timing.jl:382 [inlined]
│     [20] err(f::Any, listener::Module, modname::String, file::String, line::Any)
│        @ Requires ~/.julia/packages/Requires/Z8rfN/src/require.jl:47
│     [21] (::Vlasiator.var"#230#236")()
│        @ Vlasiator ~/.julia/packages/Requires/Z8rfN/src/require.jl:100
│     [22] withpath(f::Any, path::String)
│        @ Requires ~/.julia/packages/Requires/Z8rfN/src/require.jl:37
│     [23] (::Vlasiator.var"#229#235")()
│        @ Vlasiator ~/.julia/packages/Requires/Z8rfN/src/require.jl:99
│     [24] #invokelatest#2
│        @ ./essentials.jl:729 [inlined]
│     [25] invokelatest
│        @ ./essentials.jl:726 [inlined]
│     [26] foreach(f::typeof(Base.invokelatest), itr::Vector{Function})
│        @ Base ./abstractarray.jl:2774
│     [27] loadpkg(pkg::Base.PkgId)
│        @ Requires ~/.julia/packages/Requires/Z8rfN/src/require.jl:27
│     [28] #invokelatest#2
│        @ ./essentials.jl:729 [inlined]
│     [29] invokelatest
│        @ ./essentials.jl:726 [inlined]
│     [30] run_package_callbacks(modkey::Base.PkgId)
│        @ Base ./loading.jl:869
│     [31] _require_prelocked(uuidkey::Base.PkgId)
│        @ Base ./loading.jl:1206
│     [32] macro expansion
│        @ ./loading.jl:1180 [inlined]
│     [33] macro expansion
│        @ ./lock.jl:223 [inlined]
│     [34] require(into::Module, mod::Symbol)
│        @ Base ./loading.jl:1144
│     [35] macro expansion
│        @ ~/.julia/packages/Vlasiator/oOdeL/test/runtests.jl:380 [inlined]
│     [36] macro expansion
│        @ .../julia-1.8.2/share/julia/stdlib/v1.8/Test/src/Test.jl:1357 [inlined]
│     [37] macro expansion
│        @ ~/.julia/packages/Vlasiator/oOdeL/test/runtests.jl:380 [inlined]
│     [38] macro expansion
│        @ .../julia-1.8.2/share/julia/stdlib/v1.8/Test/src/Test.jl:1357 [inlined]
│     [39] top-level scope
│        @ ~/.julia/packages/Vlasiator/oOdeL/test/runtests.jl:39
│     [40] include(fname::String)
│        @ Base.MainInclude ./client.jl:476
│     [41] top-level scope
│        @ none:6
│     [42] eval
│        @ ./boot.jl:368 [inlined]
│     [43] exec_options(opts::Base.JLOptions)
│        @ Base ./client.jl:276
│     [44] _start()
│        @ Base ./client.jl:522
│    in expression starting at ~/.julia/packages/Vlasiator/oOdeL/src/plot/pyplot.jl:14
└ @ Requires ~/.julia/packages/Requires/Z8rfN/src/require.jl:51
PyPlot: Error During Test at ~/.julia/packages/Vlasiator/oOdeL/test/runtests.jl:379
  Got exception outside of a @test
  PyError ($(Expr(:escape, :(ccall(#= ~/.julia/packages/PyCall/ygXW2/src/pyfncall.jl:43 =# @pysym(:PyObject_Call), PyPtr, (PyPtr, PyPtr, PyPtr), o, pyargsptr, kw))))) <class 'ValueError'>
  ValueError('Illegal format string "proton/vg_rho"; two marker symbols')
    File "/usr/lib/python3/dist-packages/matplotlib/pyplot.py", line 2787, in plot
      return gca().plot(
    File "/usr/lib/python3/dist-packages/matplotlib/axes/_axes.py", line 1665, in plot
      lines = [*self._get_lines(*args, data=data, **kwargs)]
    File "/usr/lib/python3/dist-packages/matplotlib/axes/_base.py", line 225, in __call__
      yield from self._plot_args(this, kwargs)
    File "/usr/lib/python3/dist-packages/matplotlib/axes/_base.py", line 366, in _plot_args
      linestyle, marker, color = _process_plot_format(tup[-1])
    File "/usr/lib/python3/dist-packages/matplotlib/axes/_base.py", line 90, in _process_plot_format
      raise ValueError(

  Stacktrace:
    [1] pyerr_check
      @ ~/.julia/packages/PyCall/ygXW2/src/exception.jl:62 [inlined]
    [2] pyerr_check
      @ ~/.julia/packages/PyCall/ygXW2/src/exception.jl:66 [inlined]
    [3] _handle_error(msg::String)
      @ PyCall ~/.julia/packages/PyCall/ygXW2/src/exception.jl:83
    [4] macro expansion
      @ ~/.julia/packages/PyCall/ygXW2/src/exception.jl:97 [inlined]
    [5] #107
      @ ~/.julia/packages/PyCall/ygXW2/src/pyfncall.jl:43 [inlined]
    [6] disable_sigint
      @ ./c.jl:473 [inlined]
    [7] __pycall!
      @ ~/.julia/packages/PyCall/ygXW2/src/pyfncall.jl:42 [inlined]
    [8] _pycall!(ret::PyCall.PyObject, o::PyCall.PyObject, args::Tuple{MetaVLSV, String}, nargs::Int64, kw::Ptr{Nothing})
      @ PyCall ~/.julia/packages/PyCall/ygXW2/src/pyfncall.jl:29
    [9] _pycall!(ret::PyCall.PyObject, o::PyCall.PyObject, args::Tuple{MetaVLSV, String}, kwargs::Base.Pairs{Symbol, Union{}, Tuple{}, NamedTuple{(), Tuple{}}})
      @ PyCall ~/.julia/packages/PyCall/ygXW2/src/pyfncall.jl:11
   [10] pycall(::PyCall.PyObject, ::Type{PyCall.PyAny}, ::MetaVLSV, ::Vararg{Any}; kwargs::Base.Pairs{Symbol, Union{}, Tuple{}, NamedTuple{(), Tuple{}}})
      @ PyCall ~/.julia/packages/PyCall/ygXW2/src/pyfncall.jl:83
   [11] pycall(::PyCall.PyObject, ::Type{PyCall.PyAny}, ::MetaVLSV, ::Vararg{Any})
      @ PyCall ~/.julia/packages/PyCall/ygXW2/src/pyfncall.jl:83
   [12] plot(::MetaVLSV, ::Vararg{Any}; kws::Base.Pairs{Symbol, Union{}, Tuple{}, NamedTuple{(), Tuple{}}})
      @ PyPlot ~/.julia/packages/PyPlot/NsAWt/src/PyPlot.jl:194
   [13] plot(::MetaVLSV, ::Vararg{Any})
      @ PyPlot ~/.julia/packages/PyPlot/NsAWt/src/PyPlot.jl:190
   [14] macro expansion
      @ ~/.julia/packages/Vlasiator/oOdeL/test/runtests.jl:384 [inlined]
   [15] macro expansion
      @ .../julia-1.8.2/share/julia/stdlib/v1.8/Test/src/Test.jl:1357 [inlined]
   [16] macro expansion
      @ ~/.julia/packages/Vlasiator/oOdeL/test/runtests.jl:380 [inlined]
   [17] macro expansion
      @ .../julia-1.8.2/share/julia/stdlib/v1.8/Test/src/Test.jl:1357 [inlined]
   [18] top-level scope
      @ ~/.julia/packages/Vlasiator/oOdeL/test/runtests.jl:39
   [19] include(fname::String)
      @ Base.MainInclude ./client.jl:476
   [20] top-level scope
      @ none:6
   [21] eval
      @ ./boot.jl:368 [inlined]
   [22] exec_options(opts::Base.JLOptions)
      @ Base ./client.jl:276
   [23] _start()
      @ Base ./client.jl:522
WARNING: could not import Vlasiator.vdfslice into Main
Plot UI: Error During Test at ~/.julia/packages/Vlasiator/oOdeL/test/runtests.jl:472
  Got exception outside of a @test
  UndefVarError: pui not defined
  Stacktrace:
    [1] macro expansion
      @ ~/.julia/packages/Suppressor/tQJeL/src/Suppressor.jl:66 [inlined]
    [2] macro expansion
      @ ~/.julia/packages/Vlasiator/oOdeL/test/runtests.jl:475 [inlined]
    [3] macro expansion
      @ .../julia-1.8.2/share/julia/stdlib/v1.8/Test/src/Test.jl:1357 [inlined]
    [4] macro expansion
      @ ~/.julia/packages/Vlasiator/oOdeL/test/runtests.jl:473 [inlined]
    [5] macro expansion
      @ .../julia-1.8.2/share/julia/stdlib/v1.8/Test/src/Test.jl:1357 [inlined]
    [6] top-level scope
      @ ~/.julia/packages/Vlasiator/oOdeL/test/runtests.jl:39
    [7] include(fname::String)
      @ Base.MainInclude ./client.jl:476
    [8] top-level scope
      @ none:6
    [9] eval
      @ ./boot.jl:368 [inlined]
   [10] exec_options(opts::Base.JLOptions)
      @ Base ./client.jl:276
   [11] _start()
      @ Base ./client.jl:522
Test Summary:           | Pass  Error  Total   Time
Vlasiator.jl            |  114      2    116  47.1s
  Reading files         |   56            56   7.7s
  Derived variables     |   33            33   9.4s
  VLSV writing          |    1             1   0.4s
  Rotation              |    3             3   0.7s
  Curvature, Divergence |    3             3   0.6s
  Gradient              |    3             3   0.4s
  FluxFunction          |    3             3   1.6s
  Sampling              |    4             4   1.6s
  VTK                   |    3             3   2.7s
  Log                   |    1             1   1.3s
  Monitor               |    1             1   0.1s
  PyPlot                |           1      1  12.2s
  Plots                 |    3             3   3.4s
  Plot UI               |           1      1   0.3s
ERROR: LoadError: Some tests did not pass: 114 passed, 0 failed, 2 errored, 0 broken.
in expression starting at ~/.julia/packages/Vlasiator/oOdeL/test/runtests.jl:38
ERROR: Package Vlasiator errored during testing
```

Xref https://github.com/openjournals/joss-reviews/issues/4906
