### 2022-11-23 henry2004y/Vlasiator.jl#115 @ranocha
JOSS review: documentation
- [x] I am not familier with Vlasiator. Thus, it would be nice for someone like me to include a basic tutorial how to obtain such simulation data (or at least link to a Vlasiator tutorial).
- [x] It would be great to get a basic introduction to the package already in the README.md, e.g., basic installation instructions and a simple use case.
- [x] It may be confusing that the URLs in the "About" section on GitHub and in the README.md point to the development documentation, not the latest stable release (which people will usually install).
- [x] It would be nice to add a link from the documentation to the repository (to make it easier for people to see the code when they find the documentation with a search engine). For example, the first sentence on https://henry2004y.github.io/Vlasiator.jl/stable could start with "[Vlasiator.jl](https://github.com/henry2004y/Vlasiator.jl) is a data processing and analyzing tool" instead of "Data processing and analyzing tool".
- [x] The gallery at https://henry2004y.github.io/Vlasiator.jl/stable/gallery/ is nice - but it lacks links to the source code used to generate the plots.

Xref https://github.com/openjournals/joss-reviews/issues/4906
