### 2022-11-23 henry2004y/Vlasiator.jl#116 @ranocha
JOSS review: functionality
- [x] The statement "Vlasiator.jl achieves optimal serial performance for single file processing" in the paper needs to be confirmed and backed up by data.
- [x] Reading "This part has not been carefully tested so it might not work or just generate wrong results. Contributions from users are warmly welcomed!" in the documentation at https://henry2004y.github.io/Vlasiator.jl/stable/manual/#Computing-derived-quantities does not make me feel good when trying to use this package. Maybe some additional testing would be good before publishing this software in JOSS? From the JOSS guidelines at https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort: "JOSS requires that software should be feature-complete (i.e., no half-baked solutions)".

Xref https://github.com/openjournals/joss-reviews/issues/4906
