### 2022-11-23 henry2004y/Vlasiator.jl#117 @ranocha
JOSS review: software paper
- [x] How is Vlasiator.jl different from the Python package Analysator?
- [x] Vlasiator is sometimes written in normal font, sometimes in monospaced (code) font.
- [x] It would be good to cite the standard Julia reference.
- [x] The references should be checked. For example, the Zenodo entry of Plots.jl does indicate quite a few more authors.

Xref https://github.com/openjournals/joss-reviews/issues/4906
