### 2022-11-23 henry2004y/Vlasiator.jl#115 @henry2004y
Thanks for the review! I have one question regarding links to the plotting source codes: the data for generating some of the good-looking plots are either private or too large to put on a public server. How should I handle that in general? Are links to sample plotting scripts without data enough?
