### 2022-11-23 henry2004y/Vlasiator.jl#115 @ranocha
Could you maybe provide some sample data you can use as a stub (also for testing)?
