### 2022-11-23 henry2004y/Vlasiator.jl#114 @henry2004y
This is caused by an unsuccessful installation of `PyCall`, which in turn affect `PyPlot`. I did some more tests on different platforms:

| Platform | Affiliation            | Information | Status |
|----------|------------------------|-------------|-------|
| Ubuntu   | PC                     | Intel(R) i5-10210U | Passed |
| Win11    | PC                     | Intel(R) i7-10750H | Passed |
| Turso    | University of Helsinki | AMD EPYC 7452 | Passed |
| Puhti    | CSC                    | Intel(R) Xeon(R) Gold 6230 | Passed |
| Mahti    | CSC                    | AMD EPYC 7402 | Failed |

On Mahti, I see similar error messages as posted above when linking to the system loaded Python. When instead installing a fresh standalone Python via conda:

```julia
ENV["PYTHON"] = ""
Pkg.build("PyCall")
```

It almost passed all tests except one for `contour`:

```julia
PyPlot: Error During Test at /users/hongyang/.julia/packages/Vlasiator/oOdeL/test/runtests.jl:379
  Got exception outside of a @test
  PyError ($(Expr(:escape, :(ccall(#= /users/hongyang/.julia/packages/PyCall/ygXW2/src/pyfncall.jl:43 =# @pysym(:PyObject_Call), PyPtr, (PyPtr, PyPtr, PyPtr), o, pyargsptr, kw))))) <class 'ImportError'>
  ImportError("/appl/spack/v017/install-tree/gcc-8.5.0/gcc-11.2.0-zshp2k/lib64/libstdc++.so.6: version `GLIBCXX_3.4.30' not found (required by /users/hongyang/.julia/conda/3/lib/python3.10/site-packages/contourpy/_contourpy.cpython-310-x86_64-linux-gnu.so)")
    File "/users/hongyang/.julia/conda/3/lib/python3.10/site-packages/matplotlib/__init__.py", line 1423, in inner
      return func(ax, *map(sanitize_sequence, args), **kwargs)
    File "/users/hongyang/.julia/conda/3/lib/python3.10/site-packages/matplotlib/axes/_axes.py", line 6398, in contour
      contours = mcontour.QuadContourSet(self, *args, **kwargs)
    File "/users/hongyang/.julia/conda/3/lib/python3.10/site-packages/matplotlib/contour.py", line 794, in __init__
      kwargs = self._process_args(*args, **kwargs)
    File "/users/hongyang/.julia/conda/3/lib/python3.10/site-packages/matplotlib/contour.py", line 1419, in _process_args
      import contourpy
    File "/users/hongyang/.julia/conda/3/lib/python3.10/site-packages/contourpy/__init__.py", line 3, in <module>
      from ._contourpy import (
```
