### 2022-11-23 6d18f38
Add simple guide in README
