### 2022-11-23 ebb6851
Wrap into a function
