+++ docs/src/manual.md
@@ -183,9 +183,6 @@ which can also be found as keys of dictionary in [vlsvvariables.jl](https://gith
 !!! note
     In Vlasiator, the cells inside the inner boundary (which is usually a sphere/circle) are filled with zero density values. This is then used to identify the inner boundary for all other quantities. Therefore, if you are manipulating directly on data, make sure that the nonsense values inside the inner boundary are excluded. One way to do this can be found in [vlsvvariables.jl](https://github.com/henry2004y/Vlasiator.jl/blob/master/src/vlsv/vlsvvariables.jl).
 
-!!! warning
-    This part has not been carefully tested so it might not work or just generate wrong results. Contributions from users are warmly welcomed!
-
 ### Velocity space moments
 
 We can also calculate plasma moments from the saved VLSV velocity space distributions.
