### 2022-11-23 eabce3f
Differentiate the pkg and Vlasiator itself
