### 2022-11-24 henry2004y/Vlasiator.jl#116 @henry2004y
Update on the status about the derived variable calculations:

As of v0.9.32, we include tests for all listed variables. However, due to the intrinsic output options of Vlasiator, the same derived quantity may be calculated from different sources (e.g. scalar pressure from pressure tensor, or directly from VDF), or the required variables for computing others may have alternatives (e.g. current density from DCCRG B, or FSGRID B). Out current approach uses the most commonly used ones, whereas alternatives are left as future improvements. In terms of software development, this part is feature-complete; in terms of science, it is definitely not.
