### 2022-11-25 acfe8d0
Add threading for extractsat
