### 2022-11-28 henry2004y/Vlasiator.jl#116 @henry2004y
For the serial file I/O performance, we have separated the benchmarks (that was under the Log Section) to a [stand-alone section](https://henry2004y.github.io/Vlasiator.jl/dev/benchmark/) in the doc. This is used to highlight the improvement over the original Python package [Analysator](https://github.com/fmihpc/analysator). For array size larger than ~ 1MB, we can achieve ~ 3000 MB/s reading speed, which is quite close to what a typical SSD4 can get. We also realized that accurately estimating the theoretical maximum I/O performance on each platform is tricky, and is also beyond the scope of showing our current improvements over the Python implementation. Therefore, we do not show any comparison with respect to the system limits.
