### 2022-11-28 e9ecb67
Update performance comparisons
