### 2022-11-28 henry2004y/Vlasiator.jl#116 @ranocha
Thanks, this looks good! I guess some `Julia, 1st time [s]` need to be changed to `Julia, 2nd time [s]`. When that is done, all my comments above are addressed and this issue can be closed.
