+++ paper/paper.bib
@@ -1,5 +1,18 @@
+@article{julia,
+    title={Julia: A fresh approach to numerical computing},
+    author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
+    journal={SIAM {R}eview},
+    volume={59},
+    number={1},
+    pages={65--98},
+    year={2017},
+    publisher={SIAM},
+    doi={10.1137/141000671},
+    url={https://epubs.siam.org/doi/10.1137/141000671}
+}
+
 @article{palmroth2018vlasov,
-  url = {https://link.springer.com/article/10.1007/s41115-018-0003-2},
+  doi = {10.1007/s41115-018-0003-2}
   title={Vlasov methods in space physics and astrophysics},
   author={Palmroth, Minna and Ganse, Urs and Pfau-Kempf, Yann and Battarbee, Markus and Turc, Lucile and Brito, Thiago and Grandin, Maxime and Hoilijoki, Sanni and Sandroos, Arto and von Alfthan, Sebastian},
   journal={Living Reviews in Computational Astrophysics},
@@ -11,7 +24,7 @@
 }
 
 @article{von2014vlasiator,
-  url = {https://doi.org/10.1016/j.jastp.2014.08.012},
+  doi = {10.1016/j.jastp.2014.08.012},
   title={Vlasiator: First global hybrid-Vlasov simulations of Earth's foreshock and magnetosheath},
   author={Von Alfthan, S and Pokhotelov, D and Kempf, Y and Hoilijoki, S and Honkonen, I and Sandroos, A and Palmroth, M},
   journal={Journal of Atmospheric and Solar-Terrestrial Physics},
@@ -22,7 +35,7 @@
 }
 
 @article{honkonen2013parallel,
-  url = {https://doi.org/10.1016/j.cpc.2012.12.017},
+  doi = {10.1016/j.cpc.2012.12.017},
   title={Parallel grid library for rapid and flexible simulation development},
   author={Honkonen, Ilja and von Alfthan, Sebastian and Sandroos, Arto and Janhunen, Pekka and Palmroth, Minna},
   journal={Computer Physics Communications},
@@ -34,8 +47,7 @@
 }
 
 @article{pc5,
-  url = {http://www.frontiersin.org/articles/10.3389/fspas.2022.984918/abstract},
-  doi = {10.3389/fspas.2022.984918},
+  doi = {10.1016/j.cpc.2012.12.017},
   title={Magnetospheric Responses to Solar Wind Pc5 Density Fluctuations: Results from 2D Hybrid Vlasov Simulation},
   author={Zhou, Hongyang and Turc, Lucile and Pfau-Kempf, Yann and Battarbee, Markus and Tarvus, Vertti and Dubart, Maxime and George, Harriet and Cozzani, Giulia and Grandin, Maxime and Ganse, Urs and Alho, Markku and Johlander, Andreas and Suni, Jonas and Bussov, Maarja and Papadakis, Konstantinos and Horaites, Konstantinos and Zaitsev, Ivan and Tesema, Fasil and Gordeev, Evgeniy and Palmroth, Minna},
   journal={Frontiers in Astronomy and Space Sciences},
@@ -43,14 +55,15 @@
   publisher={Frontiers}
 }
 
-@misc{analysator,
-  author = {Battarbee, Markus and Akseli Hannuksela, Otto and Pfau-Kempf, Yann and von Alfthan, Sebastian and Ganse, Urs and Jarvinen, Riku and Grandin, Maxime, Lkotipal, Leo and Suni, Jonas and Alho, Markku},
-  title = {Analysator v0.9},
-  year = {2021},
-  publisher = {GitHub},
[…]
