+++ paper/paper.bib
@@ -12,7 +12,7 @@
 }
 
 @article{palmroth2018vlasov,
-  doi = {10.1007/s41115-018-0003-2}
+  doi = {10.1007/s41115-018-0003-2},
   title={Vlasov methods in space physics and astrophysics},
   author={Palmroth, Minna and Ganse, Urs and Pfau-Kempf, Yann and Battarbee, Markus and Turc, Lucile and Brito, Thiago and Grandin, Maxime and Hoilijoki, Sanni and Sandroos, Arto and von Alfthan, Sebastian},
   journal={Living Reviews in Computational Astrophysics},
