### 2022-11-28 henry2004y/Vlasiator.jl#117 @henry2004y
1. Vlasiator.jl retains all the actively used functionalities in [Analysator](https://github.com/fmihpc/analysator). Furthermore, we provide [format conversions to VTK](https://henry2004y.github.io/Vlasiator.jl/dev/manual/#Converting-to-VTK), [VLSV format modifications](https://henry2004y.github.io/Vlasiator.jl/dev/manual/#Appending-to-VLSV), more [velocity space distribution analysis methods](https://henry2004y.github.io/Vlasiator.jl/dev/manual/#Velocity-space-moments), [tracking log files](https://henry2004y.github.io/Vlasiator.jl/dev/manual/#Tracking-log-files) and more. The main features listed in the Summary Section is a superset of Analysator.
2. We have modified the text such that `Vlasiator.jl` as well as other codes/libraries are always in monospaced font.
3. We have added the reference to Julia itself in the beginning of the manuscript.
4. We have revisited the references and used consistent doi links, if possible. The Plot.jl Zenodo link has many author names in the others list, but in the [BibTeX export](https://zenodo.org/record/7368531/export/hx#.Y4SqnjNBwUE) there is only the primary author's name.
