### 2022-11-28 henry2004y/Vlasiator.jl#117 @ranocha
Everything is fine from my side if the explanations in 1. above are reflected in the paper. In this case, this issue can be closed.
