### 2022-11-28 henry2004y/Vlasiator.jl#115 @henry2004y
Currently the test data can be obtained from [vlsv_data](https://github.com/henry2004y/vlsv_data), which is mentioned in the [Log Section](https://henry2004y.github.io/Vlasiator.jl/dev/log/#Test-Data) of the documentation. These files are also used for testing. For potential Vlasiator.jl users, it is highly likely that they will first obtain some pre-existing data from private servers and then perform analysis with Vlasiator.jl. Hopefully these reference VLSV output data are enough for checking the validity of the software --- they may be changed in the future if we want higher coverage rate or new features are requested.
