### 2022-11-28 henry2004y/Vlasiator.jl#115 @ranocha
Sounds good to me. This issue can be closed once the software version in the paper is updated accordingly.
