### 2022-11-28 #4906 @tkoskela
I've read the COI policy and I may have a conflict of interest. I was employed in the same institution as the author, I left in July 2019, less than four years ago. I've never met or worked with the author. Can you please advise if I can continue the review?
