### 2022-11-28 cd668a8
Fix typo in the eq
