+++ Project.toml
@@ -24,7 +24,7 @@ WriteVTK = "64499a7a-5c06-52f2-abe2-ccb03c286192"
 [compat]
 EzXML = "1.1"
 LaTeXStrings = "1.2"
-LazyGrids = "0.4"
+LazyGrids = "0.4, 0.5"
 Parsers = "2.4"
 PyPlot = "2.9"
 RecipesBase = "1.1"
