+++ src/plot/pythonplot.jl
@@ -1,0 +1,526 @@
+# Plotting functionalities from Matplotlib.
+
+using PythonPlot
+using PythonCall: Py, pyconvert
+using REPL.TerminalMenus # Command line UI
+
+export plot, pcolormesh, vdfslice, streamplot, quiver, plotmesh
+export pui
+
+@static if pyconvert(String, matplotlib.__version__) ≥ "3.3"
+   matplotlib.rc("image", cmap="turbo") # set default colormap
+end
+
+@static if pyconvert(String, matplotlib.__version__) < "3.5"
+   matplotlib.rc("pcolor", shading="nearest") # newer version default "auto"
+end
+
+matplotlib.rc("font", size=14)
+matplotlib.rc("xtick", labelsize=10)
+matplotlib.rc("ytick", labelsize=10)
+
+
+"""
+    plot(meta::MetaVLSV, var::String, ax=nothing; comp=0, kwargs...)
+
+Plot 1D `var` from data linked to VLSV `meta`. If `ax===nothing`, plot on the current axes.
+The keyword `comp` is used to specify the component index of a vector, with 0 being the
+magnitude. Any valid keyword argument for `pyplot.plot` is accepted.
+
+# Keywords
+- `comp::Int`: the component of a vector, chosen from any valid integers.
+"""
+function PythonPlot.plot(meta::MetaVLSV, var::String, ax::Union{Py,Nothing}=nothing;
+   comp::Int=0, kwargs...)
+   ndims(meta) == 1 || error("plot only accepts 1D data!")
+
+   datafull = readvariable(meta, var)
+
+   if ndims(datafull) == 1
+      data = datafull
+   else
+      if comp == 0
+         data = sqrt.(sum(x->x^2, datafull, dims=1)) |> vec
+      else
+         data = datafull[comp,:]
+      end
+   end
+
+   x = LinRange(meta.coordmin[1], meta.coordmax[1], meta.ncells[1])
+
+   isnothing(ax) && (ax = pyplot.gca())
+
+   ax.plot(x, data; kwargs...)
+end
+
+"""
+    streamplot(meta::MetaVLSV, var::String, ax=nothing; comp::String="xy", axisunit=EARTH,
+       kwargs...)
+
+Wrapper over Matplotlib's streamplot function.
+
+# Keywords
+- `comp::String`: a subset of "xyz" in any order.
+- `axisunit::AxisUnit`: chosen from `EARTH` and `SI`.
+
+Any valid keyword argument for `pyplot.streamplot` is accepted.
+"""
+function PythonPlot.streamplot(meta::MetaVLSV, var::String, ax::Union{Py,Nothing}=nothing;
+   comp::String="xy", axisunit::AxisUnit=EARTH, kwargs...)
+
+   X, Y, v1, v2 = set_vector(meta, var, comp, axisunit)
+
+   isnothing(ax) && (ax = pyplot.gca())
+
+   streamplot(X, Y, v1, v2; kwargs...)
+end
+
+"""
+    quiver(meta::MetaVLSV, var::String, ax=nothing; comp="xy", axisunit=EARTH, stride=10,
+       kwargs...)
+
+Wrapper over Matplotlib's quiver function. If `ax===nothing`, plot on the current axes.
+
+# Keywords
+- `comp::String`: a subset of "xyz" in any order.
+- `axisunit::AxisUnit`: chosen from `EARTH` and `SI`.
+- `stride::Int`: arrow strides in number of cells.
+
+Any valid keyword argument for `pyplot.quiver` is accepted.
+"""
[…]
