+++ Project.toml
@@ -14,6 +14,7 @@ Parsers = "69de0a69-1ddd-5017-9359-2bf0b02dc9f0"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 PyPlot = "d330b81b-6aea-500a-939a-2ce795aea3ee"
 PythonCall = "6099a3de-0909-46bc-b1f4-468b9a2dfc0d"
+PythonPlot = "274fc56d-3b97-40fa-a1cd-1b4a50311bf9"
 REPL = "3fa0cd96-eef1-5676-8a61-b3b8758bbffb"
 RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
 Requires = "ae029012-a4dd-5104-9daa-d747884805df"
@@ -28,6 +29,8 @@ LaTeXStrings = "1.2"
 LazyGrids = "0.4, 0.5"
 Parsers = "2.4"
 PyPlot = "2.9"
+PythonCall = "0.9"
+PythonPlot = "1.0"
 RecipesBase = "1.1"
 Requires = "1.1"
 SnoopPrecompile = "1.0"
