### 2022-12-02 723938a
Remove direct matplotlib call
