+++ src/plot/matplotlib.jl
@@ -1,420 +1,0 @@
-# Plotting functionalities from Matplotlib through PythonCall.
-
-using PythonCall
-using REPL.TerminalMenus # Command line UI
-
-export plot, pcolormesh, pcolormeshslice, vdfslice, streamplot, quiver, plotmesh
-export pui
-export mpl, plt # common Python equivalent shorthand notations
-
-mpl = pyimport("matplotlib")
-mpl.use("tkagg") # enforce backend
-
-plt = pyimport("matplotlib.pyplot")
-
-@static if pyconvert(String, mpl.__version__) ≥ "3.3"
-   mpl.rc("image", cmap="turbo") # set default colormap
-end
-
-@static if pyconvert(String, mpl.__version__) < "3.5"
-   mpl.rc("pcolor", shading="nearest") # newer version default "auto"
-end
-
-mpl.rc("font", size=14)
-mpl.rc("xtick", labelsize=10)
-mpl.rc("ytick", labelsize=10)
-
-
-"""
-    plot(meta, var, ax=nothing; comp=0, kwargs)
-
-Plot `var` from `meta` of 1D VLSV data. If `ax===nothing`, plot on the current active axes.
-The keyword `comp` in (0, 1, 2, 3) is used to specify the component of a vector.
-"""
-function plot(meta::MetaVLSV, var, ax=nothing; comp=0, kwargs...)
-   ndims(meta) == 1 || error("plot only accepts 1D data!")
-
-   datafull = readvariable(meta, var)
-
-   if comp == 0
-      data = datafull
-   else
-      data = datafull[comp,:]
-   end
-
-   x = LinRange(meta.coordmin[1], meta.coordmax[1], meta.ncells[1])
-
-   isnothing(ax) && (ax = plt.gca())
-
-   ax.plot(x, data; kwargs...)
-end
-
-"""
-    streamplot(meta, var, ax=nothing; comp="xy", axisunit=EARTH, kwargs...)
-
-Wrapper over Matplotlib's streamplot function.
-The keyword arguments can be any valid Matplotlib arguments into streamplot.
-
-# Optional arguments
-- `comp`: a subset of "xyz" in any order.
-- `axisunit`: chosen from `EARTH` and `SI`.
-"""
-function streamplot(meta::MetaVLSV, var::AbstractString, ax=nothing;
-   comp="xy", axisunit=EARTH, kwargs...)
-
-   X, Y, v1, v2 = set_vector(meta, var, comp, axisunit)
-
-   isnothing(ax) && (ax = plt.gca())
-
-   streamplot(X, Y, v1, v2; kwargs...)
-end
-
-"""
-    quiver(meta, var, ax=nothing; comp="xy", axisunit=EARTH, stride=10, kwargs...)
-
-Wrapper over Matplotlib's quiver function. If `ax===nothing`, plot on the current active
-axes.
-The keyword arguments can be any valid Matplotlib arguments into quiver.
-
-# Optional arguments
-- `comp`: a subset of "xyz" in any order.
-- `axisunit`: chosen from `EARTH` and `SI`.
-- `stride::Integer`: arrow strides in number of cells.
-"""
-function quiver(meta::MetaVLSV, var::AbstractString, ax=nothing;
-   comp="xy", axisunit::AxisUnit=EARTH, stride::Integer=10, kwargs...)
-
-   X, Y, v1, v2 = set_vector(meta, var, comp, axisunit)
-
-   isnothing(ax) && (ax = plt.gca())
-
-   Xq,  Yq  = X[1:stride:end, 1:stride:end],  Y[1:stride:end, 1:stride:end]
-   v1q, v2q = v1[1:stride:end, 1:stride:end], v2[1:stride:end, 1:stride:end]
-
[…]
