### 2022-12-09 henry2004y/Vlasiator.jl#121 @tkoskela
JOSS review: Documentation
- [x] I think the [documentation](https://henry2004y.github.io/Vlasiator.jl/stable/benchmark/) should include instructions on how to reproduce the results. I've ran `benchmarks.jl` using `PkgBenchmarks` but it's not entirely clear to me how the results relate to what is reported in the documentation. The files `perf.jl` and `perf.py` probably should not be included in the package as they contain hard-coded file and path names that aren't part of the package.
- [x] The paper comments on how good the interoperability is with other community packages. I'm curious why you have chosen to define your own physical constants, instead of using a community package, such as [PhysicalConstants.jl](https://github.com/JuliaPhysics/PhysicalConstants.jl)
- [x] The link to [repo](https://github.com/henry2004y/Vlasiator.jl/tree/master/src/examples) in https://github.com/henry2004y/Vlasiator.jl/blob/feda666f96ed87fb2928221680b2b5cd5473106b/docs/src/manual.md?plain=1#L233 is broken

https://github.com/openjournals/joss-reviews/issues/4906
