### 2022-12-09 henry2004y/Vlasiator.jl#122 @tkoskela
JOSS review: features
- [x] The usage examples seem to assume you have access to data from Vlasiator. It's difficult to comment on the features because the examples won't run without data. I noticed the benchmarks pull data from an artifact, could you provide data for the examples as well?

ref https://github.com/openjournals/joss-reviews/issues/4906
