### 2022-12-09 henry2004y/Vlasiator.jl#123 @tkoskela
JOSS review: Software paper
- [x] You probably mean *tera*bytes here https://github.com/henry2004y/Vlasiator.jl/blob/702cba67e794397b07e6683833f48fa18b11390e/paper/paper.md?plain=1#L43
- [x] Could you provide links to the repositories of Vlasiator and Analysator in the references?

ref https://github.com/openjournals/joss-reviews/issues/4906
