### 2022-12-10 henry2004y/Vlasiator.jl#121 @henry2004y
Thanks for the review!

1. The link is fixed.
2. PhysicalConstants.jl is a nice collection of constants, but it come naturally with the dependencies on Unitful.jl which is something we haven't tested carefully. Having predefined units is good, but it unavoidably comes at a cost of performance. Also, in order to make computations convenient and consistent, we may have to define my own unit system for commonly used units in our community, which is also not fully tested. Thus we decided to define the constants inside the package (which is currently only ~ 10 lines). Currently based on this [list](https://juliahub.com/ui/Packages/PhysicalConstants/ykpMf/0.2.3?page=2), a few of the astrophysics packages depend on it.  In the future, we may make a switch if PhysicalConstants.jl proves to be a nice addition.
3. About the benchmarks between Julia and Python, the most tricky part, as the reviewers also pointed out in #122 and #115, is that VLSV format files generated from Vlasiator are either large or non-public. In order to show the comparisons at real work situations, our current solution is to run `perf.jl` and `perf.py` on private data on the server with different sizes. The [test data](https://henry2004y.github.io/Vlasiator.jl/stable/log/#Test-Data) for this package contains 3 VLSV files which are intently kept small. If we use those for the benchmarks, then only reading DCCRG grid variables at very small array sizes can be shown, which may not be ideal. Our current plan is to update `perf.jl` and `perf.py` and make better descriptions to the benchmark results.
