### 2022-12-10 henry2004y/Vlasiator.jl#122 @henry2004y
The way to obtain test data is described [here](https://henry2004y.github.io/Vlasiator.jl/stable/log/#Test-Data). We noticed that the other reviewer also had comments on this, which indicates that this part should be moved to another proper section in the documentation. WIP.
