### 2022-12-10 henry2004y/Vlasiator.jl#123 @henry2004y
We have modified the manuscript accordingly.
