### 2022-12-15 eafeb73
Distinguish adiabatic and polytropic
