### 2022-12-15 henry2004y/Vlasiator.jl#123 @tkoskela
This looks good
