### 2022-12-15 henry2004y/Vlasiator.jl#121 @tkoskela
Apologies, I had not looked under `Log` for instructions on how to run the benchmarks. I think renaming it to FAQ makes it a bit more obvious, but I would also put a link to it from the Benchmarks page. Probably that's part of your plans in making better descriptions of the benchmark results.

I understand that the size of the data files may prevent you from distributing them as part of the package. I don't really have an universal solution for that, I would suggest looking at services your institute or HPC center may offer for data storage and distribution. I'm perhaps a bit confused about the issue of data being non-public. If there is no public data, it's difficult to see the benefits of an open source processing tool. I understand you might want to keep unpublished data private, but you should have something published that can be used to run the benchmarks on.

I would suggest at least clearly documenting the benchmarks that only make sense to run on large datasets that may not be available as part of the package, and providing data for the benchmarks where you can get meaningful results within the limits of what you can store on GitHub.
