### 2022-12-15 henry2004y/Vlasiator.jl#122 @tkoskela
Again, sorry, the Log section confused me. Could you put a link to it in the "Loading VLSV data" section?
