### 2022-12-15 henry2004y/Vlasiator.jl#121 @henry2004y
Thanks for the suggestions! Upon inquiring our colleagues, we now find some open-access Vlasiator data. We will update the benchmarks and add the links to the data.
