+++ benchmark/perf.jl
@@ -1,66 +1,82 @@
 # Script for generating benchmark results in
 # https://henry2004y.github.io/Vlasiator.jl/dev/benchmark/
-#
-# Note: this script is not intended for execution as a whole: each benchmark needs to be
-# executed independently after importing all the required packages.
 
-using Vlasiator
-using Vlasiator: RE
+using Vlasiator, PyPlot
 using BenchmarkTools
-using Glob
 
-## Reading DCCRG variables
+ENV["MPLBACKEND"]="agg" # no GUI for Matplotlib
 
-file1 = "bulk.singleprecision.vlsv" # 80 B
-file2 = "bulk.0000003.vlsv"         # 900 KB
-file3 = "bulk1.0001000.vlsv"        # 32 MB
-# Select file
-file = file1
-# Load metadata
-@time meta = load(file);
-@time meta = load(file);
-@time meta = load(file);
-# Select variable name
-var = "proton/vg_rho";
-# Read variable, sorted
-@time rho = meta[var];
-@time rho = meta[var];
-# Read variable, unsorted
-@time rho_unsorted = readvariable($meta, $var, false);
-@time rho_unsorted = readvariable($meta, $var, false);
+files = ["1d_single.vlsv", "bulk.2d.vlsv", "2d_double.vlsv", "2d_AFC.vlsv", "3d_EGI.vlsv"]
+versions = [5,5,5,4,5]
 
-## Plotting with PyPlot
-# Log color scale is used for comparison with Analysator.
-
-# 2D density contour on a uniform mesh
-filename = "bulk.0000501.vlsv"
-meta = load(filename)
-@time pcolormesh(meta, "rho", colorscale=Log)
-close()
-@time pcolormesh(meta, "rho", colorscale=Log)
-
-# 2D density slices from 3D AMR mesh
-filename = "bulk1.0001000.vlsv"
-meta = load(filename)
-@time pcolormesh(meta, "proton/vg_rho", colorscale=Log)
-close()
-@time pcolormesh(meta, "proton/vg_rho", colorscale=Log)
-
-## Virtual satellite tracking at a static location
-
-# 3D EGI data on Turso, University of Helsinki
-dir = "/wrk-vakka/group/spacephysics/vlasiator/3D/EGI/bulk/dense_cold_hall1e5_afterRestart374/"
-
-filenames = glob("bulk1*.vlsv", dir)
-
-var = "proton/vg_rho"
-loc = [12, 0, 0] .* RE
-
-println("Number of files: ", length(filenames))
-# Static cell ID
-id = let
-   meta = load(filenames[1])
-   getcell(meta, loc)
+# Download test files if not found in the current path
+for i in eachindex(files)
+   if isfile(files[i])
+      @info "Benchmark file $(files[i]) found..."
+      continue
+   elseif i == 4
+      url_base = "https://a3s.fi/swift/v1/AUTH_81f1cd490d494224880ea77e4f98490d/vlasiator-2d-afc/"
+      filename = "production_halfres/bulk.0000000.vlsv"
+      file_origin = basename(filename)
+      url = joinpath(url_base, filename)
+      @info "Downloading 1.8G test file..."
+      run(`curl -o $file_new $url`)
+      run(`mv $(file_origin) $file`)
+   elseif i in (1,2,3)
+      @info "Downloading test files..."
+      run(`curl -o testdata.tar.gz https://raw.githubusercontent.com/henry2004y/vlsv_data/master/testdata.tar.gz`)
+      run(`curl -o 1d_single.vlsv https://raw.githubusercontent.com/henry2004y/vlsv_data/master/1d_single.vlsv`)
+      run(`curl -o 2d_double.vlsv https://raw.githubusercontent.com/henry2004y/vlsv_data/master/2d_double.vlsv`)
[…]

[…]
