+++ src/vlsv/vlsvreader.jl
@@ -257,8 +257,8 @@ function load(file::AbstractString)
       nodeblockid)
 
    cellid = getcellid(fid, n.var)
-
-   cellindex = sortperm(cellid)
+   #TODO: check the new sorting algorithms in Julia 1.9!
+   cellindex = sortperm(cellid, alg=MergeSort)
 
    ncells, block_size, coordmin, coordmax = readmesh(fid, footer)
 
