+++ docs/src/benchmark.md
@@ -50,9 +50,9 @@ Access to the public data can be found from [vlsv_data](https://github.com/henry
 | 4          | 0.5       | 12.3       | 9.4   |
 
 * Plotting 2D density slices from a 3D AMR mesh
-| Size     | Julia [s][^4] | Python [s] | Ratio |
-|:---------|:-------------:|:----------:|:-----:|
-| 30.5 MiB | 0.5           | 4.6        | 10    |
+| File Index | Julia [s][^4] | Python [s] | Ratio |
+|:-----------|:-------------:|:----------:|:-----:|
+| 5          | 0.5           | 4.6        | 10    |
 
 [^4]: The first time execution will be slower due to JIT compilation (which is excluded in the timing here). This is known as "Time-To-First-X" in the Julia community.
 
