### 2022-12-16 029cbc0
Modify AMR plotting stats
