### 2022-12-16 henry2004y/Vlasiator.jl#121 @henry2004y
In Vlasiator.jl v0.9.34, we have updated `perf.jl`, `perf.py` as well as the benchmark results. `perf.jl` and `perf.py` can be run directly to produce a major portion of the benchmarks listed in the doc. One of the files (file index 5) is private, while others are all publicly accessible. We have also fixed some issues in comparing the performance between Python/Numpy and Julia, which should now give fairer results.
