### 2023-01-04 d0ec441
Use map for type conversion
