### 2023-01-04 33a0efb
Add more type annotations
