### 2023-01-05 cc160ca
Refactor getnearestcelwithvdf
