### 2023-01-05 f7ecec7
Replace sum with loop
