### 2023-01-06 a22238e
Add type annotations to avoid mem alloc
