+++ README.md
@@ -3,7 +3,7 @@
 <p align="center">
   <img src="docs/src/figures/logo_fancy_black.png" height="200"><br>
   <a href="https://github.com/henry2004y/Vlasiator.jl/actions">
-    <img src="https://img.shields.io/github/workflow/status/henry2004y/Vlasiator.jl/CI">
+    <img src="https://img.shields.io/github/actions/workflow/status/henry2004y/Vlasiator.jl/CI.yml?branch=master">
   </a>
   <a href="https://codecov.io/gh/henry2004y/Vlasiator.jl">
     <img src="https://img.shields.io/codecov/c/github/henry2004y/Vlasiator.jl">
