### 2023-01-11 #4906 @ranocha
From my side, only https://github.com/henry2004y/Vlasiator.jl/issues/114 needs to be fixed to check the remaining items on my list
