### 2023-01-11 henry2004y/Vlasiator.jl#114 @henry2004y
I tried again on my previous failed platform Mahti with Julia1.8.5 and Vlasiator v0.9.36, and now it has passed:

<details><summary>Mahti</summary>

```julia
(@v1.8) pkg> st
Status `~/.julia/environments/v1.8/Project.toml`
  [6e4b80f9] BenchmarkTools v1.3.2
  [438e738f] PyCall v1.95.1
  [d330b81b] PyPlot v2.11.0
  [7d2ba682] Vlasiator v0.9.36

(@v1.8) pkg> test Vlasiator
     Testing Vlasiator
      Status `/local_scratch/hongyang/jl_qDiQp9/Project.toml`
  [8f5d6c58] EzXML v1.1.0
  [b964fa9f] LaTeXStrings v1.3.0
  [7031d0ef] LazyGrids v0.5.0
  [69de0a69] Parsers v2.5.2
  [d330b81b] PyPlot v2.11.0
  [3cdcf5f2] RecipesBase v1.3.3
  [ae029012] Requires v1.3.0
  [66db9d55] SnoopPrecompile v1.0.3
  [90137ffa] StaticArrays v1.5.12
  [fd094767] Suppressor v0.2.1
  [7d2ba682] Vlasiator v0.9.36
  [64499a7a] WriteVTK v1.16.0
  [ade2ca70] Dates `@stdlib/Dates`
  [4af54fe1] LazyArtifacts `@stdlib/LazyArtifacts`
  [37e2e46d] LinearAlgebra `@stdlib/LinearAlgebra`
  [a63ad114] Mmap `@stdlib/Mmap`
  [de0858da] Printf `@stdlib/Printf`
  [3fa0cd96] REPL `@stdlib/REPL`
  [ea8e919c] SHA v0.7.0 `@stdlib/SHA`
  [10745b16] Statistics `@stdlib/Statistics`
  [8dfed614] Test `@stdlib/Test`
      Status `/local_scratch/hongyang/jl_qDiQp9/Manifest.toml`
  [944b1d66] CodecZlib v0.7.0
  [3da002f7] ColorTypes v0.11.4
  [5ae59095] Colors v0.12.10
  [8f4d0f93] Conda v1.7.0
  [8f5d6c58] EzXML v1.1.0
  [1a297f60] FillArrays v0.13.6
  [53c48c17] FixedPointNumbers v0.8.4
  [692b3bcd] JLLWrappers v1.4.1
  [682c06a0] JSON v0.21.3
  [b964fa9f] LaTeXStrings v1.3.0
  [7031d0ef] LazyGrids v0.5.0
  [9c8b4983] LightXML v0.9.0
  [1914dd2f] MacroTools v0.5.10
  [69de0a69] Parsers v2.5.2
  [21216c6a] Preferences v1.3.0
  [438e738f] PyCall v1.95.1
  [d330b81b] PyPlot v2.11.0
  [3cdcf5f2] RecipesBase v1.3.3
  [189a3867] Reexport v1.2.2
  [ae029012] Requires v1.3.0
  [66db9d55] SnoopPrecompile v1.0.3
  [90137ffa] StaticArrays v1.5.12
  [1e83bf80] StaticArraysCore v1.4.0
  [fd094767] Suppressor v0.2.1
  [3bb67fe8] TranscodingStreams v0.9.11
  [81def892] VersionParsing v1.3.0
  [7d2ba682] Vlasiator v0.9.36
  [64499a7a] WriteVTK v1.16.0
  [94ce4f54] Libiconv_jll v1.16.1+2
  [02c8fc9c] XML2_jll v2.10.3+0
  [0dad84c5] ArgTools v1.1.1 `@stdlib/ArgTools`
  [56f22d72] Artifacts `@stdlib/Artifacts`
  [2a0f44e3] Base64 `@stdlib/Base64`
  [ade2ca70] Dates `@stdlib/Dates`
  [f43a241f] Downloads v1.6.0 `@stdlib/Downloads`
  [7b1f6079] FileWatching `@stdlib/FileWatching`
  [b77e0a4c] InteractiveUtils `@stdlib/InteractiveUtils`
  [4af54fe1] LazyArtifacts `@stdlib/LazyArtifacts`
  [b27032c2] LibCURL v0.6.3 `@stdlib/LibCURL`
  [76f85450] LibGit2 `@stdlib/LibGit2`
  [8f399da3] Libdl `@stdlib/Libdl`
  [37e2e46d] LinearAlgebra `@stdlib/LinearAlgebra`
  [56ddb016] Logging `@stdlib/Logging`
  [d6f4376e] Markdown `@stdlib/Markdown`
  [a63ad114] Mmap `@stdlib/Mmap`
  [ca575930] NetworkOptions v1.2.0 `@stdlib/NetworkOptions`
  [44cfe95a] Pkg v1.8.0 `@stdlib/Pkg`
  [de0858da] Printf `@stdlib/Printf`
  [3fa0cd96] REPL `@stdlib/REPL`
  [9a3f8284] Random `@stdlib/Random`
  [ea8e919c] SHA v0.7.0 `@stdlib/SHA`
  [9e88b42a] Serialization `@stdlib/Serialization`
  [6462fe0b] Sockets `@stdlib/Sockets`
  [2f01184e] SparseArrays `@stdlib/SparseArrays`
  [10745b16] Statistics `@stdlib/Statistics`
  [fa267f1f] TOML v1.0.0 `@stdlib/TOML`
  [a4e569a6] Tar v1.10.1 `@stdlib/Tar`
  [8dfed614] Test `@stdlib/Test`
  [cf7118a7] UUIDs `@stdlib/UUIDs`
  [4ec0a83e] Unicode `@stdlib/Unicode`
  [e66e0078] CompilerSupportLibraries_jll v1.0.1+0 `@stdlib/CompilerSupportLibraries_jll`
  [deac9b47] LibCURL_jll v7.84.0+0 `@stdlib/LibCURL_jll`
  [29816b5a] LibSSH2_jll v1.10.2+0 `@stdlib/LibSSH2_jll`
  [c8ffd9c3] MbedTLS_jll v2.28.0+0 `@stdlib/MbedTLS_jll`
  [14a3606d] MozillaCACerts_jll v2022.2.1 `@stdlib/MozillaCACerts_jll`
  [4536629a] OpenBLAS_jll v0.3.20+0 `@stdlib/OpenBLAS_jll`
  [83775a58] Zlib_jll v1.2.12+3 `@stdlib/Zlib_jll`
  [8e850b90] libblastrampoline_jll v5.1.1+0 `@stdlib/libblastrampoline_jll`
  [8e850ede] nghttp2_jll v1.48.0+0 `@stdlib/nghttp2_jll`
  [3f19e933] p7zip_jll v17.4.0+0 `@stdlib/p7zip_jll`
Precompiling project...
  1 dependency successfully precompiled in 1 seconds. 35 already precompiled.
     Testing Running tests...
Test Summary: | Pass  Total     Time
Vlasiator.jl  |  139    139  1m02.1s
     Testing Vlasiator tests passed
```

Maybe it's due to some recent fix in PyCall. Can you check if this works on your test machine @ranocha?
