### 2023-01-11 henry2004y/Vlasiator.jl#114 @ranocha
Seem to work on my system, too.
