### 2023-01-22 e44af63
Search for the artifact path instead of symlink
