### 2023-01-23 69c0f36
Add DCCRG vector benchmark
