+++ benchmark/benchmarks.jl
@@ -19,7 +19,8 @@ SUITE["read"] = BenchmarkGroup(["variable"])
 file = joinpath(directory, files[2])
 meta = load(file)
 SUITE["read"]["meta"] = @benchmarkable load($file)
-SUITE["read"]["DCCRG"] = @benchmarkable readvariable($meta, "proton/vg_rho")
+SUITE["read"]["DCCRG_scalar"] = @benchmarkable readvariable($meta, "proton/vg_rho")
+SUITE["read"]["DCCRG_vector"] = @benchmarkable readvariable($meta, "proton/vg_v")
 SUITE["read"]["DCCRG_derived"] = @benchmarkable readvariable($meta, "Vperp")
 ids = collect(100:110)
 SUITE["read"]["DCCRG_select_small"] =
