+++ src/Vlasiator.jl
@@ -31,8 +31,8 @@ export
    hasvariable, hasparameter, hasname,
    # vlsvutility
    getcell, getslicecell, getlevel, refineslice, getcellcoordinates, getvcellcoordinates,
-   getchildren, getparent, isparent, getsiblings,
-   getcellinline, getnearestcellwithvdf, getcellwithvdf, getmaxwellianity,
+   getchildren, getparent, isparent, getsiblings, getcellinline, getnearestcellwithvdf,
+   getcellwithvdf, getmaxwellianity, getKLdivergence,
    getdensity, getvelocity, getpressure, getheatfluxvector, write_vtk, write_vlsv, issame,
    # plot helper
    SI, EARTH, Log, Linear, SymLog,
