### 2023-01-24 080f013
Revert "Try again the symbolic link solution"
This reverts commit 22adfaf5d7b43d85c40a6ec5b6d43c70df35d5bf.
