### 2023-01-24 f254526
Revert "Search for the artifact path instead of symlink"
This reverts commit e44af63de66306b68cf660d28a6199c5d4a0f736.
