### 2023-01-26 henry2004y/Vlasiator.jl#122 @tkoskela
Sorry for taking so long to review. I can run the examples and the Pluto notebook looks very cool! I'm happy to close this, I'd like to note that the `Plotting` section of the notebook does not work for me with the `bulk.amr.vlsv` file, maybe that is to be expected since you talk about cell IDs without AMR in the markdown.

This is the error I get

```
UndefVarError: origin not defined

macro expansion@plots.jl:49[inlined]
apply_recipe(::AbstractDict{Symbol, Any}, ::Vlasiator.MetaVLSV, ::AbstractString)@RecipesBase.jl:289
_process_userrecipes!(::Any, ::Any, ::Any)@user_recipe.jl:36
recipe_pipeline!(::Any, ::Any, ::Any)@RecipesPipeline.jl:70
_plot!(::Plots.Plot, ::Any, ::Any)@plot.jl:208
#plot#139@plot.jl:91[inlined]
var"#contourf#457"(::Base.Pairs{Symbol, V, Tuple{Vararg{Symbol, N}}, NamedTuple{names, T}} where {V, N, names, T<:Tuple{Vararg{Any, N}}}, ::typeof(Plots.contourf), ::Any, ::Vararg{Any})@RecipesBase.jl:410
top-level scope@Local: 3[inlined]
```
