### 2023-01-26 henry2004y/Vlasiator.jl#121 @tkoskela
I would still like to ask for a bit clearer documentation on Benchmarks. With some guesswork I can link some of the numbers reported in https://henry2004y.github.io/Vlasiator.jl/dev/benchmark/ to the output of `benchmarkpkg(Vlasiator)`, for example file index 2 of "loading meta data" is probably `"meta" => Trial(310.480 μs)`. If you could describe on the Benchmarks page which numbers I can expect to extract from running benchmarkpkg on the data you provide, and how, that would make me happy.
