### 2023-01-26 henry2004y/Vlasiator.jl#121 @tkoskela
I would also note that I ran into an error
```
ERROR: LoadError: Cannot locate '(Julia)Artifacts.toml' file when attempting to use artifact 'testdata' in 'Main'
```
that seems to be reported in https://github.com/JuliaLang/IJulia.jl/issues/1060. The solution in the first comment of the issue fixed it for me
