### 2023-01-26 #4906 @tkoskela
Sorry for the delay @pdebuyl. I tried out the examples and benchmarks using the new documentation and had one more question for the author. I'll try to respond quickly
