### 2023-01-26 henry2004y/Vlasiator.jl#121 @henry2004y
I ran this issue as well and indeed the first comment fixed it. However, for some reason it didn't work for the pkgbenchmark that I added recently. In CI, my current workaround is to manually create a symbolic link under the benchmark folder:
https://github.com/henry2004y/Vlasiator.jl/blob/19b203b46d9d602c9463057652a8c21756082169/.github/workflows/benchmark.yml#L26

Then this works for both CI and local pkgbenchmark.
