### 2023-01-27 b057b39
Fix Plots.jl 3D slices
