+++ examples/demo_pluto.jl
@@ -1,5 +1,5 @@
 ### A Pluto.jl notebook ###
-# v0.19.4
+# v0.19.20
 
 using Markdown
 using InteractiveUtils
@@ -15,7 +15,6 @@ macro bind(def, element)
 end
 
 # ╔═╡ 9ba5e632-e78a-44f8-b80b-1cb5a3cf8042
-# ╠═╡ show_logs = false
 using Vlasiator, PlutoUI, Plots, Downloads, Tar, CodecZlib
 
 # ╔═╡ 7c6789f2-688d-425a-8f70-4d465ea00104
@@ -137,12 +136,14 @@ The plasma moments computed from VDFs should be pretty close to the direct outpu
 
 # ╔═╡ 7228f7d6-c5fc-41c6-a955-16fdb0ae14f2
 md"## Plotting
-Visualization is always fun! Let's start with plotting the cell IDs. Without adaptive mesh refinement (AMR), cell IDs start with 1 and are always continuous."
+Visualization is always fun! Let's start with plotting the cell IDs. Without adaptive mesh refinement (AMR), cell IDs start with 1 and are always continuous; with AMR, cell IDs may not be continuous due to the mesh generation of DCCRG."
 
 # ╔═╡ 7af713c0-ef87-4eb2-9cf8-ebdb95f65437
 let
-   if ndims(meta) == 2 || ndims(meta) == 3
-      contourf(meta, "CellID", aspect_ratio=:equal, c=:turbo, xlims=(-8,30), size=(400,400))
+   if ndims(meta) == 2
+      contourf(meta, "CellID", aspect_ratio=:equal, c=:turbo, xlims=(-8,30), size=(600,400))
+   elseif ndims(meta) == 3
+	   contourf(meta, "CellID", aspect_ratio=:equal, c=:turbo, xlims=(-13,12), ylims=(-6,6), size=(600,400))
    elseif ndims(meta) == 1
 	  plot(meta, "CellID", size=(600,300))
    end
@@ -158,8 +159,10 @@ end
 # ╔═╡ 8612f699-83db-4d02-ae16-c5c33c04e09e
 # ╠═╡ show_logs = false
 begin
-   if ndims(meta) == 2 || ndims(meta) == 3
+   if ndims(meta) == 2
       heatmap(meta, "proton/vg_blocks", aspect_ratio=:equal, c=:turbo, size=(400,580))
+   elseif ndims(meta) == 3
+      heatmap(meta, "proton/vg_blocks", aspect_ratio=:equal, c=:turbo, size=(600,400))
    elseif ndims(meta) == 1
 	  plot(meta, "proton/vg_rho")
    end
@@ -180,18 +183,19 @@ Tar = "a4e569a6-e804-4fa4-b0f3-eef7a1d5b13e"
 Vlasiator = "7d2ba682-ad6e-4e20-80d9-3f2d4a610bb4"
 
 [compat]
-CodecZlib = "~0.7.0"
-Plots = "~1.29.0"
-PlutoUI = "~0.7.38"
-Vlasiator = "~0.9.20"
+CodecZlib = "~0.7.1"
+Plots = "~1.38.2"
+PlutoUI = "~0.7.49"
+Vlasiator = "~0.9.37"
 """
 
 # ╔═╡ 00000000-0000-0000-0000-000000000002
 PLUTO_MANIFEST_TOML_CONTENTS = """
[…]
