+++ .github/workflows/benchmark.yml
@@ -15,7 +15,7 @@ jobs:
   Benchmark:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
       - uses: julia-actions/setup-julia@v1
         with:
           version: 1
