### 2023-01-27 henry2004y/Vlasiator.jl#121 @tkoskela
I was running `benchmarks.jl`, thank you for pointing that out! I can verify the benchmark data, so I am happy to close the issue. Labeling the benchmark suites in `perf.jl` to match the documentation a bit better could be helpful.
