### 2023-01-27 henry2004y/Vlasiator.jl#121 @henry2004y
Yeah this is indeed kind of confusing, because currently PkgBenchmarks requires the file benchmarks.jl so I cannot use it for comparing against Python with exactly the same tasks and larger datasets. But we will find a clearer (and automatic) way of doing this!
