+++ src/vlsv/vlsvreader.jl
@@ -62,7 +62,8 @@ function Base.show(io::IO, meta::MetaVLSV)
    print(io, "File       : ")
    printstyled(io, meta.name, '\n'; color=:cyan, underline=true)
    print(io, "Time       : ")
-   printstyled(io, round(meta.time, digits=2), '\n'; color=:cyan)
+   t = @sprintf "%.4e s" meta.time
+   printstyled(io, t, '\n'; color=:cyan)
    print(io, "Dimension  : ")
    printstyled(io, ndims(meta), '\n'; color=:yellow)
    print(io, "Max AMR lvl: ")
