### 2023-02-11 03e90eb
Merge sort is still faster in Julia 1.9-beta4
