+++ .github/workflows/CI.yml
@@ -6,6 +6,7 @@ on:
     paths-ignore:
       - 'README.md'
       - 'examples/**'
+      - 'install/**'
   push:
     branches:
       - master
@@ -14,6 +15,7 @@ on:
       - 'README.md'
       - 'examples/**'
       - 'benchmark/**'
+      - 'install/**'
 env:
   JULIA_NUM_THREADS: 1
 jobs:
