### 2023-02-13 8918f0f
Handle hasvdf
