### 2023-02-13 576acab
Support VLSV with no vspace info (#138)
* Handle #137
