+++ src/Vlasiator.jl
@@ -1,6 +1,5 @@
 module Vlasiator
 
-using Requires
 using StaticArrays: SVector, @SVector, SMatrix, @SMatrix
 using Printf: @sprintf
 using LinearAlgebra: ×, dot, ⋅, norm, normalize, normalize!
@@ -13,6 +12,9 @@ using LaTeXStrings
 using Dates
 using Parsers
 using SnoopPrecompile
+if !isdefined(Base, :get_extension)
+   using Requires
+end
 
 include("utility/rotation.jl")
 include("utility/log.jl")
@@ -49,12 +51,14 @@ export
    meta = load(initfile)
 end
 
-function __init__()
-   @require PyPlot="d330b81b-6aea-500a-939a-2ce795aea3ee" begin
-      include("plot/pyplot.jl")
-   end
-   @require Plots="91a5bcdd-55d7-5caf-9e0b-520d859cae80" begin
-      include("plot/plots.jl")
+@static if !isdefined(Base, :get_extension)
+   function __init__()
+      @require PyPlot="d330b81b-6aea-500a-939a-2ce795aea3ee" begin
+         include("../ext/PyPlotExt.jl")
+      end
+      @require Plots="91a5bcdd-55d7-5caf-9e0b-520d859cae80" begin
+         include("../ext/PlotsExt.jl")
+      end
    end
 end
 
