+++ ext/PlotsExt.jl
@@ -1,6 +1,15 @@
-# Using user recipes from Plots.
+"Using user recipes from Plots."
+module PlotsExt
 
-using RecipesBase, Printf
+using Vlasiator
+using Vlasiator: AxisUnit, ColorScale, PlotArgs,
+   set_args, prep2d, get_axis, _fillinnerBC!, set_lim
+if isdefined(Base, :get_extension)
+   using RecipesBase
+else
+   using ..RecipesBase
+end
+using Printf
 
 # Build a recipe which acts on a custom type.
 @recipe function f(meta::MetaVLSV, var::AbstractString;
@@ -96,5 +105,7 @@ end
       bins --> r1, r2
       v1, v2
    end
+
+end
 
 end
