### 2023-02-21 #4906 @pdebuyl
Hello @henry2004y

1. Zhou 2022 has the DOI of Honkonen 2013, can you fix it
2. Can you provide direct download for the examples in the gallery ? https://henry2004y.github.io/Vlasiator.jl/stable/gallery/ The code refers to filenames directly without pointing to the source of the data.

(sorry for the delay)
