+++ paper/paper.bib
@@ -47,7 +47,7 @@
 }
 
 @article{pc5,
-  doi = {10.1016/j.cpc.2012.12.017},
+  doi = {10.3389/fspas.2022.984918},
   title={Magnetospheric Responses to Solar Wind Pc5 Density Fluctuations: Results from 2D Hybrid Vlasov Simulation},
   author={Zhou, Hongyang and Turc, Lucile and Pfau-Kempf, Yann and Battarbee, Markus and Tarvus, Vertti and Dubart, Maxime and George, Harriet and Cozzani, Giulia and Grandin, Maxime and Ganse, Urs and Alho, Markku and Johlander, Andreas and Suni, Jonas and Bussov, Maarja and Papadakis, Konstantinos and Horaites, Konstantinos and Zaitsev, Ivan and Tesema, Fasil and Gordeev, Evgeniy and Palmroth, Minna},
   journal={Frontiers in Astronomy and Space Sciences},
