### 2023-02-21 #4906 @henry2004y
Thanks for pointing out the DOI error. I have fixed it in my joss-paper branch.

As for the gallery download links, many of the demo plots are made with private data. Should I remake those to use public data, or add more detailed explanations about how to obtain the data and add references to the previously published papers, etc.?
