+++ Project.toml
@@ -31,6 +31,7 @@ julia = "1.7"
 
 [extras]
 LazyArtifacts = "4af54fe1-eca0-43a8-85a7-787d91b784e3"
+RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
 SHA = "ea8e919c-243c-51af-8825-aaa63cd721ce"
 Suppressor = "fd094767-a336-5f1f-9728-57cf17d0bbfb"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
