+++ src/Vlasiator.jl
@@ -12,9 +12,6 @@ using LaTeXStrings
 using Dates
 using Parsers
 using SnoopPrecompile
-if !isdefined(Base, :get_extension)
-   using Requires: @require
-end
 
 include("utility/rotation.jl")
 include("utility/log.jl")
@@ -25,6 +22,7 @@ include("vlsv/vlsvreader.jl")
 include("vlsv/vlsvvariables.jl")
 include("vlsv/vlsvutility.jl")
 include("utility/plot.jl")
+include("utility/plotrecipe.jl")
 
 export
    # vlsvreader
@@ -49,17 +47,6 @@ export
 @precompile_all_calls begin
    initfile = joinpath(@__DIR__, "../test/init.vlsv")
    meta = load(initfile)
-end
-
-@static if !isdefined(Base, :get_extension)
-   function __init__()
-      @require PyPlot="d330b81b-6aea-500a-939a-2ce795aea3ee" begin
-         include("../ext/PyPlotExt.jl")
-      end
-      @require Plots="91a5bcdd-55d7-5caf-9e0b-520d859cae80" begin
-         include("../ext/PlotsExt.jl")
-      end
-   end
 end
 
 end
