+++ test/runtests.jl
@@ -1,6 +1,6 @@
 using Vlasiator, LaTeXStrings, SHA, LazyArtifacts
 using Suppressor: @capture_out, @capture_err, @suppress_out
-using REPL.TerminalMenus
+using RecipesBase
 using Test
 
 group = get(ENV, "TEST_GROUP", :all) |> Symbol
@@ -15,24 +15,6 @@ function nanmaximum(x::AbstractArray{T}) where T<:AbstractFloat
       end
    end
    result
-end
-
-function simulate_input(keys...)
-   keydict = Dict(:up => "\e[A",
-                  :down => "\e[B",
-                  :enter => "\r")
-
-   new_stdin = Base.BufferStream()
-   for key in keys
-      if isa(key, Symbol)
-         write(new_stdin, keydict[key])
-      else
-         write(new_stdin, "$key")
-      end
-   end
-   TerminalMenus.terminal.in_stream = new_stdin
-
-   return
 end
 
 @testset "Vlasiator.jl" begin
@@ -378,83 +360,7 @@ end
    end
 
    if group in (:plot, :all)
-      @testset "PyPlot" begin
-         using PyPlot
-         pygui(false)
-         # 1D
-         meta = meta1
-         line = plot(meta, "proton/vg_rho")[1]
-         @test line.get_ydata() == meta["proton/vg_rho"]
-         line = plot(meta, "proton/vg_v", comp=3)[1]
-         @test line.get_ydata() == meta["proton/vg_v"][3,:]
-         centers = plotmesh(meta, projection="y")
-         points = centers.get_offsets()
-         @test size(points) == (10, 2)
-         fig = plt.figure()
-         ax = fig.add_subplot(projection="3d")
-         centers = plotmesh(meta, ax; projection="3d")
-         points = centers.get_offsets() # only 2D from 3D coords, might be improved
-         @test size(points) == (10, 2)
-         plt.clf()
-
-         @test_throws ArgumentError pcolormesh(meta, "proton/vg_rho")
-
-         loc = [2.0, 0.0, 0.0]
-         v = vdfslice(meta, loc).get_array()
-         @test v[786] == 238.24398578141802
-         @test_throws ArgumentError vdfslice(meta, loc, species="helium")
-         output = @capture_err begin
-            v = vdfslice(meta, loc; slicetype=:bperp).get_array()
-         end
-         @test v[786] == 4.02741885708042e-10
-
-         output = @capture_err begin
-            vdfslice(meta, loc; verbose=true)
-         end
-         @test startswith(output, "[ Info:")
-
-         # 2D
-         meta = meta2
-         v = pcolormesh(meta, "proton/vg_rho").get_array()
-         @test v[end-2] == 999535.8f0 && length(v) == 6300
-         v = pcolormesh(meta, "proton/vg_rho", extent=[0.,1.,0.,2.]).get_array()
-         @test v[end] == 0.0 && length(v) == 6
-         v = pcolormesh(meta, "proton/vg_rho";
-            extent=[-2e7, 2e7, -2e7, 2e7], axisunit=SI, colorscale=Log).get_array()
-         @test v[end] == 1.00022675f6 && length(v) == 100
-         v = pcolormesh(meta, "fg_b").get_array()
-         @test v[1] == 3.0058909f-9
-         v = pcolormesh(meta, "proton/vg_v", comp=:x, colorscale=SymLog).get_array()
-         @test v[2] == -699935.2f0
-         v = contour(meta, "proton/vg_rho").get_array()
-         @test length(v) == 8 && v[end] == 5.6e6
-         v = contourf(meta, "proton/vg_rho").get_array()
[…]
