+++ examples/demo_lineextracting_mp_pyplot.jl
@@ -1,150 +1,0 @@
-# Multiple variables alone a line across multiple frames, multi-process version.
-# Note: this script can be made faster by updating the data without deleting the lines.
-#
-# Usage:
-#   julia -p $ncores demo_vars_extracting_mp.jl
-#
-# Hongyang Zhou, hyzhou@umich.edu
-
-using Distributed, ParallelDataTransfer, Glob
-using Vlasiator: RE
-@everywhere using Vlasiator, PyPlot, Printf
-
-@everywhere function init_figure(x1, x2)
-   fig, axs = plt.subplots(4, 1; num=myid(),
-      figsize=(10, 15), sharex=true, constrained_layout=true)
-
-   axs[end].set_xlim(x1, x2)
-
-   axs[1].set_ylim(ρmin, ρmax)
-   axs[2].set_ylim(vmin, vmax)
-   axs[3].set_ylim(pmin, pmax)
-   axs[4].set_ylim(bmin, bmax)
-   for ax in axs
-      ax.xaxis.set_minor_locator(matplotlib.ticker.AutoMinorLocator())
-      ax.yaxis.set_minor_locator(matplotlib.ticker.AutoMinorLocator())
-      ax.grid(true)
-   end
-
-   axs[end].set_xlabel("x [RE]"; fontsize)
-   axs[1].set_ylabel("Density [amu/cc]"; fontsize)
-   axs[2].set_ylabel("Velocity [km/s]"; fontsize)
-   axs[3].set_ylabel("Pressure [nPa]"; fontsize)
-   axs[4].set_ylabel("Magnetic field [nT]"; fontsize)
-   return fig, axs
-end
-
-@everywhere function process(axs, file, cellids)
-   isfile("out/"*file[end-8:end-5]*".png") && return
-
-   println("file = $file")
-   meta = load(file)
-
-   p_extract = readvariable(meta, "vg_pressure", cellids) .* 1e9 |> vec # [nPa]
-   rho_extract = readvariable(meta, "proton/vg_rho", cellids) |> vec
-   v_extract = readvariable(meta, "proton/vg_v", cellids)
-   vmag2_extract = sum(x -> x*x, v_extract, dims=1) |> vec
-   pram_extract = rho_extract .* Vlasiator.mᵢ .* vmag2_extract .* 1e9 # [nPa]
-
-   bz = readvariable(meta, "vg_b_vol", cellids)[3,:] .* 1e9 #[nT]
-
-   imagnetopause_ = findfirst(<(0.0), bz)
-
-   axs[1].plot(loc, rho_extract ./ 1e6, label="Proton density", color="#1f77b4")
-   vl1 = axs[1].vlines(loc[imagnetopause_], ρmin, ρmax;
-      colors="r", linestyle="dashed", alpha=0.5)
-
-   axs[2].plot(loc, v_extract[1,:] ./ 1e3, label="Vx", color="#1f77b4")
-   vl2 = axs[2].vlines(loc[imagnetopause_], vmin, vmax;
-      colors="r", linestyle="dashed", alpha=0.5)
-
-   axs[3].plot(loc, pram_extract, label="Ram", color="#1f77b4")
-   axs[3].plot(loc, p_extract, label="Thermal", color="#ff7f0e")
-   vl3 = axs[3].vlines(loc[imagnetopause_], pmin, pmax;
-      colors="r", linestyle="dashed", alpha=0.5)
-
-   axs[4].plot(loc, bz, label="Bz", color="#1f77b4")
-   hl4 = axs[4].hlines(0.0, loc[1], loc[end], colors="k", linestyle="dashed", alpha=0.2)
[…]
