+++ examples/demo_wave_satellite_mt.jl
@@ -1,230 +1,0 @@
-# Identification of EMIC and mirror modes in two frequency bands from single satellite data.
-#
-# EMIC: δB⟂ vs. δv_⟂
-# Mirror: δB∥ vs. δn
-#
-# The moving box window size is decided empirically.
-#
-# Usage:
-#   julia -t 4 demo_wave_satellite_mt.jl
-# or
-#   JULIA_NUM_THREADS=4 julia demo_wave_satellite_mt.jl
-#
-# Hongyang Zhou, hyzhou@umich.edu
-
-using Statistics: mean
-using LinearAlgebra: ⋅, normalize!, norm
-using DSP, PyPlot, Glob
-using Vlasiator
-using Vlasiator: μ₀, mᵢ, RE
-using Polyester
-
-"Extract variables at `loc` from VLSV `files`."
-function extract_vars(files, loc)
-   nfiles = length(files)
-
-   t = zeros(Float32, nfiles)
-   n = zeros(Float32, nfiles)
-   v = zeros(Float32, 3, nfiles)
-   b = zeros(Float32, 3, nfiles)
-   e = zeros(Float32, 3, nfiles)
-
-   id = load(files[1]) do meta
-      getcell(meta, loc)
-   end
-
-   # Extract data from each frame
-   @batch for i = eachindex(files)
-      meta = load(files[i])
-      t[i] = meta.time
-      n[i] = readvariable(meta, "proton/vg_rho", id)[1]
-      v[:,i] = readvariable(meta, "proton/vg_v", id)
-      b[:,i] = readvariable(meta, "vg_b_vol", id)
-      e[:,i] = readvariable(meta, "vg_e_vol", id)
-   end
-
-   t, n, v, e, b
-end
-
-function moving_average(g::AbstractVector{<:AbstractFloat}, n::Int)
-   nbackward = div(n,2)
-   nforward = isodd(n) ? nbackward : nbackward - 1
-   len = length(g)
-   g_avg = similar(g)
-   @inbounds @batch for i = 1:len
-      lo = max(1, i - nbackward)
-      hi = min(len, i + nforward)
-      g_avg[i] = mean(@view g[lo:hi])
-   end
-   g_avg
-end
-
-function moving_average(g::AbstractMatrix{<:AbstractFloat}, n::Int)
-   nbackward = div(n,2)
-   nforward = isodd(n) ? nbackward : nbackward - 1
-   len = size(g,2)
-   g_avg = similar(g)
-   for icomp = axes(g,1)
-      @inbounds @batch for i = 1:len
-         lo = max(1, i - nbackward)
-         hi = min(len, i + nforward)
-         g_avg[icomp,i] = mean(@view g[icomp, lo:hi])
-      end
-   end
-   g_avg
-end
-
-function detrend(v::AbstractVector{<:AbstractFloat}; nbox=length(v)÷12)
-   v̄ = moving_average(v, nbox)
-   dv = v .- v̄
-end
-
-function detrend(v::AbstractMatrix{<:AbstractFloat}; nbox=size(v,2)÷12)
-   v̄ = similar(v)
-   for i in axes(v,1)
-      v̄[i,:] = @views moving_average(v[i,:], nbox)
-   end
-   dv = v .- v̄
-end
-
-function align_yaxis(ax1, ax2)
-   y_lims = [ax.get_ylim() for ax in [ax1, ax2]]
-
-   # normalize both axes
-   y_mags = ntuple(i -> y_lims[i][2] - y_lims[i][1], Val(2))
-   y_lims_normalized = ntuple(i -> y_lims[i] ./ y_mags[i], Val(2))
-
[…]

[…]
