+++ docs/src/manual.md
@@ -20,12 +20,12 @@ A bunch of physical constants are predefined in [Vlasiator.jl](https://github.co
 
 ## Loading VLSV data
 
-A quick way to obtain small test data can be found in [F&Q](log.md#Test-Data). Larger open-access data can be found from the references in [Vlasiator publications](https://www.helsinki.fi/en/researchgroups/vlasiator/publications-and-presentations), e.g. [Takahashi+ 2020](https://a3s.fi/swift/v1/AUTH_81f1cd490d494224880ea77e4f98490d/vlasiator-2d-afc/).
+A quick way to obtain small test data can be found in [F&Q](log.md#Test-Data). Larger open-access data can be found from the references in [Vlasiator publications](https://www.helsinki.fi/en/researchgroups/vlasiator/publications-and-presentations), e.g. [Takahashi+ 2020](https://a3s.fi/swift/v1/AUTH_81f1cd490d494224880ea77e4f98490d/vlasiator-2d-afc/). In this tutorial we are using the 2D test file `bulk.2d.vlsv` from [vlsv_data](https://github.com/henry2004y/vlsv_data) if not specified.
 
 - Read meta data
 
 ```julia
-file = "bulk.0000004.vlsv"
+file = "bulk.2d.vlsv"
 meta = load(file)
 ```
 
@@ -52,9 +52,9 @@ A list of utility functions has been implemented for checking variable status. S
 - Read variable
 
 ```julia
-data = meta["proton/vg_rho"]
+data = meta["CellID"]
 # Or equivalently
-data = readvariable(meta, "proton/vg_rho")
+data = readvariable(meta, "CellID")
 ```
 
 The variable reading is designed for cells, which takes cell ID(s) as the 3rd argument if specified. The same interface works for both [DCCRG](https://github.com/fmihpc/dccrg) grid (for storing cell centered quantities like plasma moments) and [FS](https://github.com/fmihpc/fsgrid) grid (for storing field solver related quantities on a uniform high resolution mesh) variables. By default the returned DCCRG variable array is sorted by cell IDs. If in any case you want the original unsorted version as being stored in the file, use `readvariable(meta, var, false)`.
@@ -64,7 +64,7 @@ The variable reading is designed for cells, which takes cell ID(s) as the 3rd ar
 ```julia
 loc = [2.0, 0.0, 0.0]
 id = getcell(meta, loc)
-readvariable(meta, "proton/vg_rho", id)
+readvariable(meta, "CellID", id)
 ```
 
 - Get variable along a line between two points
@@ -80,6 +80,9 @@ var_extract = readvariable(meta, "VA", cellids)
 - Extract variable at a static cell ID from a sequence of files under the same grid
 
 ```julia
+files = ["bulk.2d.vlsv"]
+var = "CellID"
+id = 1
 extractsat(files, var, id)
 ```
 
@@ -190,6 +193,8 @@ which can also be found as keys of dictionary in [vlsvvariables.jl](https://gith
 We can also calculate plasma moments from the saved VLSV velocity space distributions.
 
 ```julia
+meta = load("bulk.1d.vlsv")
+cellid = 5
 # VDF cell indexes and values, with sparsity
 vcellids, vcellf = readvcells(meta, cellid; species="proton")
 
@@ -252,13 +257,13 @@ All the functions with identical names as in Matplotlib accept all possible keyw
 - Scalar colored contour from 2D simulation
 
[…]
