### 2023-03-06 #4906 @henry2004y
The [manual](https://henry2004y.github.io/Vlasiator.jl/stable/manual/#Loading-VLSV-data) is now updated in the latest stable version, with better instructions to fetch small test data. All the demos assume the usage of `bulk.2d.vlsv`, if not specified. While going through all the demos I realized that some of those were not up-to-date. I am considering shifting to Literate.jl for automatic checking of the demos, but for now it is not implemented. @pdebuyl
