### 2023-03-09 63579d5
Skip saving vtkGhostType if not required
