### 2023-03-10 870f493
Update virtual sat script
