### 2023-03-20 #4906 @pdebuyl
- The project contains benchmarks, which you use to make the performance claims of vlasiator.jl. Please refer to them (even briefly) in the article
- Please check the capitalization of Zhou 2022's title.
- I made a pull request for minor style edits. Please accept the PR if you agree or comment else.

Sorry for the delay and thanks for the update of the user manual. It was now easier for me to find the suitable data files for the user manual.
