+++ paper/paper.bib
@@ -68,7 +68,7 @@
 
 @misc{ressac,
   author = {Turc, Lucile and Zhou, Hongyang},
-  title = {Response of near-Earth Space to changing SolAr wind Conditions (RESSAC)},
+  title = {Response of near-Earth Space to changing solar wind Conditions (RESSAC)},
   year = {2020},
   publisher = {University of Helsinki},
   journal = {University of Helsinki},
