+++ paper/paper.bib
@@ -23,17 +23,6 @@
   publisher={Springer}
 }
 
-@article{von2014vlasiator,
-  doi = {10.1016/j.jastp.2014.08.012},
-  title={Vlasiator: First global hybrid-Vlasov simulations of Earth's foreshock and magnetosheath},
-  author={Von Alfthan, S and Pokhotelov, D and Kempf, Y and Hoilijoki, S and Honkonen, I and Sandroos, A and Palmroth, M},
-  journal={Journal of Atmospheric and Solar-Terrestrial Physics},
-  volume={120},
-  pages={24--35},
-  year={2014},
-  publisher={Elsevier}
-}
-
 @article{honkonen2013parallel,
   doi = {10.1016/j.cpc.2012.12.017},
   title={Parallel grid library for rapid and flexible simulation development},
@@ -68,7 +57,7 @@
 
 @misc{ressac,
   author = {Turc, Lucile and Zhou, Hongyang},
-  title = {Response of near-Earth Space to changing solar wind Conditions (RESSAC)},
+  title = {{Response of near-Earth Space to changing SolAr wind Conditions (RESSAC)}},
   year = {2020},
   publisher = {University of Helsinki},
   journal = {University of Helsinki},
@@ -77,7 +66,7 @@
 
 @article{makie,
   doi = {10.21105/joss.03349},
-  title={Makie. jl: Flexible high-performance data visualization for Julia},
+  title={Makie.jl: Flexible high-performance data visualization for Julia},
   author={Danisch, Simon and Krumbiegel, Julius},
   journal={Journal of Open Source Software},
   volume={6},
