### 2023-03-21 e078de6
Remove project journal
