+++ paper/paper.bib
@@ -60,7 +60,6 @@
   title = {{Response of near-Earth Space to changing SolAr wind Conditions (RESSAC)}},
   year = {2020},
   publisher = {University of Helsinki},
-  journal = {University of Helsinki},
   url = {https://researchportal.helsinki.fi/en/projects/response-of-near-earth-space-to-changing-solar-wind-conditions}
 }
 
