### 2023-03-21 #4906 @pdebuyl
Thank you @henry2004y I have no further requests.

Can you
1. Make a tagged release
2. archive it on zenodo (make sure to use same title and authors as JOSS article)
3. provide the doi of the zenodo archive here
