### 2023-03-22 fe822c7
Use VlasiatorPyPlot in the examples
