+++ docs/examples/visualization/demo_vdf_pyplot.md
@@ -10,7 +10,7 @@
 
 This demo shows how to plot the phase space density (i.e. velocity distribution function) near a given spatial location.
 ```julia
-using Vlasiator, PyPlot
+using Vlasiator, PyPlot, VlasiatorPyPlot
 using Vlasiator: RE # Earth radius [m]
 
 function main()
