+++ docs/examples/visualization/demo_field_tracing_2d.md
@@ -10,7 +10,6 @@
 This demo shows how to trace streamlines in a 2D uniform field.
 ```julia
 using PyPlot, FieldTracer, Vlasiator, VlasiatorPyPlot
-
 using Vlasiator: RE # Earth radius, [m]
 
 function main()
