+++ docs/examples/visualization/demo_B_fieldline_pyplot.md
@@ -10,7 +10,6 @@
 This demo shows how to plot field lines with handpicked seeds.
 ```julia
 using Vlasiator, PyPlot, VlasiatorPyPlot, FieldTracer
-
 using Vlasiator: RE # Earth radius, [m]
 
 function main()
