+++ .github/workflows/PyPlotCI.yml
@@ -1,4 +1,4 @@
-name: CI
+name: CIPyPlot
 on:
   pull_request:
     paths:
