### 2023-03-22 10a1bec
Remove empty name step
