+++ .github/workflows/CI.yml
@@ -34,7 +34,6 @@ jobs:
         arch:
           - x64
     steps:
-      - name: Install dependencies
       - uses: actions/checkout@v3
       - uses: julia-actions/setup-julia@v1
         with:
