+++ .github/workflows/PyPlotCI.yml
@@ -50,8 +50,17 @@ jobs:
             ${{ runner.os }}-test-${{ env.cache-name }}-
             ${{ runner.os }}-test-
             ${{ runner.os }}-
-      - uses: julia-actions/julia-buildpkg@v1
-      - uses: julia-actions/julia-runtest@v1
+      - name: Build package
+        shell: julia --project=monorepo {0}
+        run: |
+          using Pkg;
+          # dev mono repo versions
+          pkg"dev . ./VlasiatorPyPlot"
+      - name: Run the tests
+        continue-on-error: true
+        run: |
+          julia --project=monorepo -e 'using Pkg; Pkg.test("VlasiatorPyPlot", coverage=true)'
+          && echo "TESTS_SUCCESSFUL=true" >> $GITHUB_ENV
       - uses: julia-actions/julia-processcoverage@v1
       - uses: codecov/codecov-action@v3
         with:
