+++ .github/workflows/PyPlotCI.yml
@@ -58,7 +58,7 @@ jobs:
           pkg"dev . ./PyPlot"
       - name: Run the tests
         continue-on-error: true
-        run: |
+        run: >
           julia --project=monorepo -e 'using Pkg; Pkg.test("VlasiatorPyPlot", coverage=true)'
           && echo "TESTS_SUCCESSFUL=true" >> $GITHUB_ENV
       - uses: julia-actions/julia-processcoverage@v1
