### 2023-03-22 573d912
Drop Requires and separate PyPlot into a new pkg (#139)
* Separate PyPlot into a new pkg

* Update docs

* Update CIs
