+++ PyPlot/src/VlasiatorPyPlot.jl
@@ -1,6 +1,12 @@
-# Plotting functionalities from Matplotlib.
+"Plotting functionalities from Matplotlib."
+module VlasiatorPyPlot
 
+using Vlasiator
+using Vlasiator: AxisUnit, ColorScale, PlotArgs,
+   set_args, prep2d, prep2dslice, get_axis, _fillinnerBC!, _getdim2d, set_lim, prep_vdf,
+   SVector
 using PyPlot
+
 import PyPlot.PyCall: PyObject
 using REPL.TerminalMenus # Command line UI
 
@@ -524,3 +530,5 @@ function pui(meta::MetaVLSV; suppress_output::Bool=false)
 end
 
 pui(file::AbstractString) = file |> load |> plot_ui
+
+end
