+++ src/Vlasiator.jl
@@ -1,6 +1,5 @@
 module Vlasiator
 
-using Requires
 using StaticArrays: SVector, @SVector, SMatrix, @SMatrix
 using Printf: @sprintf
 using LinearAlgebra: ×, dot, ⋅, norm, normalize, normalize!
@@ -23,6 +22,7 @@ include("vlsv/vlsvreader.jl")
 include("vlsv/vlsvvariables.jl")
 include("vlsv/vlsvutility.jl")
 include("utility/plot.jl")
+include("utility/plotrecipe.jl")
 
 export
    # vlsvreader
@@ -47,15 +47,6 @@ export
 @precompile_all_calls begin
    initfile = joinpath(@__DIR__, "../test/init.vlsv")
    meta = load(initfile)
-end
-
-function __init__()
-   @require PyPlot="d330b81b-6aea-500a-939a-2ce795aea3ee" begin
-      include("plot/pyplot.jl")
-   end
-   @require Plots="91a5bcdd-55d7-5caf-9e0b-520d859cae80" begin
-      include("plot/plots.jl")
-   end
 end
 
 end
