+++ PyPlot/src/VlasiatorPyPlot.jl
@@ -1,11 +1,12 @@
 "Plotting functionalities from Matplotlib."
 module VlasiatorPyPlot
 
-using Vlasiator
+using Reexport
+@reexport using Vlasiator
 using Vlasiator: AxisUnit, ColorScale, PlotArgs,
    set_args, prep2d, prep2dslice, get_axis, _fillinnerBC!, _getdim2d, set_lim, prep_vdf,
    SVector
-using PyPlot
+@reexport using PyPlot
 
 import PyPlot.PyCall: PyObject
 using REPL.TerminalMenus # Command line UI
