+++ docs/examples/visualization/demo_vdfs_pyplot.md
@@ -10,7 +10,7 @@
 This demo shows how to plot velocity distribution functions along a line.
 
 ```julia
-using Vlasiator, PyPlot, VlasiatorPyPlot, JLD2
+using VlasiatorPyPlot, JLD2
 using Vlasiator: RE # Earth radius [m]
 
 function main()
