+++ PyPlot/Project.toml
@@ -13,3 +13,4 @@ Vlasiator = "7d2ba682-ad6e-4e20-80d9-3f2d4a610bb4"
 PyPlot = "2.9"
 Reexport = "1.2"
 Vlasiator = "0.10"
+julia = "1.7"
