### 2023-03-22 0262610
Set default style in init
