### 2023-03-23 7ca85ac
Add missing axes to streamplot
