### 2023-03-23 13599c4
Export set_colorbar from VlasiatorPyPlot
