### 2023-03-31 #4906 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
